chore: 🤖 fix failing tests
