+++ .changeset/empty-owls-add.md
@@ -1,0 +1,5 @@
+---
+'@lottiefiles/dotlottie-web': minor
+---
+
+refactor: 💡 dotlottie-rs wasm bindings integration
