+++ .changeset/pre.json
@@ -1,0 +1,13 @@
+{
+  "mode": "pre",
+  "tag": "next",
+  "initialVersions": {
+    "dotlottie-vue-example": "0.0.0",
+    "dotlottie-wc-example": "0.0.0",
+    "@lottiefiles/dotlottie-react": "0.2.0",
+    "@lottiefiles/dotlottie-vue": "0.1.1",
+    "@lottiefiles/dotlottie-wc": "0.0.4",
+    "@lottiefiles/dotlottie-web": "0.12.0"
+  },
+  "changesets": []
+}
