+++ packages/react/README.md
@@ -78,7 +78,7 @@ The `DotLottieReactProps` extends the `HTMLCanvasElement` Props and accepts all 
 | `src`                   | string                                 |          | undefined             | URL to the animation data (`.json` or `.lottie`).                                                                                                                                                                                                  |   |
 | `speed`                 | number                                 |          | 1                     | Animation playback speed. 1 is regular speed.                                                                                                                                                                                                      |   |
 | `data`                  | string \| ArrayBuffer                  |          | undefined             | Animation data provided either as a Lottie JSON string or as an ArrayBuffer for .lottie animations.                                                                                                                                                |   |
-| `mode`                  | string                                 |          | "forward"             | Animation play mode. Accepts "forward", "reverse", "bounce", "bounce-reverse".                                                                                                                                                                     |   |
+| `mode`                  | string                                 |          | "forward"             | Animation play mode. Accepts "forward", "reverse", "bounce", "reverse-bounce".                                                                                                                                                                     |   |
 | `backgroundColor`       | string                                 |          | undefined             | Background color of the canvas. Accepts 6-digit or 8-digit hex color string (e.g., "#000000", "#000000FF"),                                                                                                                                        |   |
 | `segments`              | \[number, number]                      |          | \[0, totalFrames - 1] | Animation segments. Accepts an array of two numbers, where the first number is the start frame and the second number is the end frame.                                                                                                             |   |
 | `renderConfig`          | RenderConfig                           |          | `{}`                  | Configuration for rendering the animation.                                                                                                                                                                                                         |   |
