--- packages/web/.dockerignore
@@ -1,1 +1,0 @@
-node_modules
