--- packages/web/docker_build_wasm.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker --log-level=debug build -t $0 . && docker run --rm -v $(pwd):/package $0
