+++ packages/web/index.html
@@ -6,18 +6,20 @@
         <title>Document</title>
     </head>
     <body>
-        <canvas width="200" height="200"></canvas>
+        <canvas style="width: 300px; height: 300px"></canvas>
         <button id="play">play</button>
         <button id="pause">pause</button>
         <button id="stop">stop</button>
         <button id="freeze">freeze</button>
         <button id="unfreeze">unfreeze</button>
+        <button id="go-to">50%</button>
         <button id="speed-up">+ speed</button>
         <button id="speed-down">- speed</button>
+        <button id="load">load</button>
         <input id="progress" type="range" min="0" max="100" defaultValue="0" step="0.01" />
 
         <script type="module">
-            import { DotLottie } from './dist/dotlottie-rs.js';
+            import { DotLottie } from './dist/index.js';
 
             const canvas = document.querySelector('canvas');
 
@@ -29,8 +31,10 @@
             const progress = document.querySelector('#progress');
             const speedUpBtn = document.querySelector('#speed-up');
             const speedDownBtn = document.querySelector('#speed-down');
+            const goToBtn = document.querySelector('#go-to');
+            const loadBtn = document.querySelector('#load');
 
-            DotLottie.setWasmUrl('./dist/core/dotlottie-player.wasm');
+            DotLottie.setWasmUrl('./dist/dotlottie-player.wasm');
 
             const response = await fetch('https://lottie.host/5525262b-4e57-4f0a-8103-cfdaa7c8969e/VCYIkooYX8.json');
             const data = await response.text();
@@ -38,13 +42,30 @@
             const dotLottie = new DotLottie({
                 canvas,
                 autoplay: true,
-                data,
+                // data,
                 loop: true,
                 useFrameInterpolation: true,
+                src: 'https://lottie.host/5525262b-4e57-4f0a-8103-cfdaa7c8969e/VCYIkooYX8.json',
             });
 
             dotLottie.addEventListener('frame', ({ currentFrame }) => {
                 progress.value = (currentFrame / dotLottie.totalFrames) * 100;
+            });
+
+            progress.addEventListener('mousedown', () => {
+                dotLottie.pause();
+            });
+
+            progress.addEventListener('mouseup', () => {
+                dotLottie.play();
+            });
+
+            progress.addEventListener('change', () => {
+                dotLottie.setFrame(dotLottie.totalFrames * (progress.value / 100));
+            });
+
+            goToBtn.addEventListener('click', () => {
+                dotLottie.setFrame(dotLottie.totalFrames * 0.5);
             });
 
             playBtn.addEventListener('click', () => {
@@ -73,6 +94,14 @@
 
             speedDownBtn.addEventListener('click', () => {
                 dotLottie.setSpeed(dotLottie.speed - 0.1);
+            });
+
+            loadBtn.addEventListener('click', async () => {
+                dotLottie.load({
+                    src: 'https://lottie.host/5525262b-4e57-4f0a-8103-cfdaa7c8969e/VCYIkooYX8.json',
+                    autoplay: true,
+                    loop: false,
+                });
             });
         </script>
     </body>
