--- packages/web/src/core/dotlottie-player.mjs
@@ -1,3311 +1,0 @@
-var createDotLottiePlayerModule = (() => {
-  var _scriptDir = typeof document !== 'undefined' && document.currentScript ? document.currentScript.src : undefined;
-
-  return function (moduleArg = {}) {
-    var Module = moduleArg;
-
-    var readyPromiseResolve, readyPromiseReject;
-
-    Module['ready'] = new Promise((resolve, reject) => {
-      readyPromiseResolve = resolve;
-      readyPromiseReject = reject;
-    });
-
-    var moduleOverrides = Object.assign({}, Module);
-
-    var arguments_ = [];
-
-    var thisProgram = './this.program';
-
-    var quit_ = (status, toThrow) => {
-      throw toThrow;
-    };
-
-    var ENVIRONMENT_IS_WEB = true;
-
-    var ENVIRONMENT_IS_WORKER = false;
-
-    var scriptDirectory = '';
-
-    function locateFile(path) {
-      if (Module['locateFile']) {
-        return Module['locateFile'](path, scriptDirectory);
-      }
-      return scriptDirectory + path;
-    }
-
-    var read_, readAsync, readBinary;
-
-    if (ENVIRONMENT_IS_WEB || ENVIRONMENT_IS_WORKER) {
-      if (ENVIRONMENT_IS_WORKER) {
-        scriptDirectory = self.location.href;
-      } else if (typeof document != 'undefined' && document.currentScript) {
-        scriptDirectory = document.currentScript.src;
-      }
-      if (_scriptDir) {
-        scriptDirectory = _scriptDir;
-      }
-      if (scriptDirectory.indexOf('blob:') !== 0) {
-        scriptDirectory = scriptDirectory.substr(0, scriptDirectory.replace(/[?#].*/, '').lastIndexOf('/') + 1);
-      } else {
-        scriptDirectory = '';
-      }
-      {
-        read_ = (url) => {
-          var xhr = new XMLHttpRequest();
-          xhr.open('GET', url, false);
-          xhr.send(null);
-          return xhr.responseText;
-        };
-        if (ENVIRONMENT_IS_WORKER) {
-          readBinary = (url) => {
-            var xhr = new XMLHttpRequest();
-            xhr.open('GET', url, false);
-            xhr.responseType = 'arraybuffer';
-            xhr.send(null);
-            return new Uint8Array(/** @type{!ArrayBuffer} */ (xhr.response));
-          };
-        }
-        readAsync = (url, onload, onerror) => {
-          var xhr = new XMLHttpRequest();
-          xhr.open('GET', url, true);
-          xhr.responseType = 'arraybuffer';
-          xhr.onload = () => {
-            if (xhr.status == 200 || (xhr.status == 0 && xhr.response)) {
-              onload(xhr.response);
-              return;
-            }
-            onerror();
-          };
-          xhr.onerror = onerror;
-          xhr.send(null);
-        };
-      }
-    } else {
-    }
-
-    var out = Module['print'] || console.log.bind(console);
-
-    var err = Module['printErr'] || console.error.bind(console);
-
-    Object.assign(Module, moduleOverrides);
-
-    moduleOverrides = null;
-
-    if (Module['arguments']) arguments_ = Module['arguments'];
-
-    if (Module['thisProgram']) thisProgram = Module['thisProgram'];
-
-    if (Module['quit']) quit_ = Module['quit'];
-
-    var wasmBinary;
-
-    if (Module['wasmBinary']) wasmBinary = Module['wasmBinary'];
-
-    if (typeof WebAssembly != 'object') {
-      abort('no native wasm support detected');
-    }
-
-    var wasmMemory;
-
-    var ABORT = false;
-
-    var EXITSTATUS;
-
-    var /** @type {!Int8Array} */ HEAP8,
-      /** @type {!Uint8Array} */ HEAPU8,
-      /** @type {!Int16Array} */ HEAP16,
-      /** @type {!Uint16Array} */ HEAPU16,
-      /** @type {!Int32Array} */ HEAP32,
-      /** @type {!Uint32Array} */ HEAPU32,
-      /** @type {!Float32Array} */ HEAPF32,
-      /** @type {!Float64Array} */ HEAPF64;
-
-    function updateMemoryViews() {
-      var b = wasmMemory.buffer;
-      Module['HEAP8'] = HEAP8 = new Int8Array(b);
-      Module['HEAP16'] = HEAP16 = new Int16Array(b);
-      Module['HEAPU8'] = HEAPU8 = new Uint8Array(b);
-      Module['HEAPU16'] = HEAPU16 = new Uint16Array(b);
-      Module['HEAP32'] = HEAP32 = new Int32Array(b);
-      Module['HEAPU32'] = HEAPU32 = new Uint32Array(b);
-      Module['HEAPF32'] = HEAPF32 = new Float32Array(b);
-      Module['HEAPF64'] = HEAPF64 = new Float64Array(b);
-    }
-
-    var __ATPRERUN__ = [];
-
-    var __ATINIT__ = [];
-
-    var __ATPOSTRUN__ = [];
-
-    var runtimeInitialized = false;
-
-    function preRun() {
-      if (Module['preRun']) {
-        if (typeof Module['preRun'] == 'function') Module['preRun'] = [Module['preRun']];
-        while (Module['preRun'].length) {
-          addOnPreRun(Module['preRun'].shift());
-        }
-      }
-      callRuntimeCallbacks(__ATPRERUN__);
-    }
-
-    function initRuntime() {
-      runtimeInitialized = true;
-      callRuntimeCallbacks(__ATINIT__);
-    }
-
-    function postRun() {
-      if (Module['postRun']) {
-        if (typeof Module['postRun'] == 'function') Module['postRun'] = [Module['postRun']];
-        while (Module['postRun'].length) {
-          addOnPostRun(Module['postRun'].shift());
-        }
-      }
-      callRuntimeCallbacks(__ATPOSTRUN__);
-    }
-
-    function addOnPreRun(cb) {
-      __ATPRERUN__.unshift(cb);
-    }
-
-    function addOnInit(cb) {
-      __ATINIT__.unshift(cb);
-    }
-
-    function addOnPostRun(cb) {
-      __ATPOSTRUN__.unshift(cb);
-    }
-
-    var runDependencies = 0;
-
-    var runDependencyWatcher = null;
-
-    var dependenciesFulfilled = null;
-
-    function addRunDependency(id) {
-      runDependencies++;
-      Module['monitorRunDependencies']?.(runDependencies);
-    }
-
-    function removeRunDependency(id) {
-      runDependencies--;
-      Module['monitorRunDependencies']?.(runDependencies);
-      if (runDependencies == 0) {
-        if (runDependencyWatcher !== null) {
-          clearInterval(runDependencyWatcher);
-          runDependencyWatcher = null;
-        }
-        if (dependenciesFulfilled) {
-          var callback = dependenciesFulfilled;
-          dependenciesFulfilled = null;
-          callback();
-        }
-      }
-    }
-
-    /** @param {string|number=} what */ function abort(what) {
-      Module['onAbort']?.(what);
-      what = 'Aborted(' + what + ')';
-      err(what);
-      ABORT = true;
-      EXITSTATUS = 1;
-      what += '. Build with -sASSERTIONS for more info.';
-      /** @suppress {checkTypes} */ var e = new WebAssembly.RuntimeError(what);
-      readyPromiseReject(e);
-      throw e;
-    }
-
-    var dataURIPrefix = 'data:application/octet-stream;base64,';
-
-    /**
-     * Indicates whether filename is a base64 data URI.
-     * @noinline
-     */ var isDataURI = (filename) => filename.startsWith(dataURIPrefix);
-
-    var wasmBinaryFile;
-
-    wasmBinaryFile = 'DotLottiePlayer.wasm';
-
-    if (!isDataURI(wasmBinaryFile)) {
-      wasmBinaryFile = locateFile(wasmBinaryFile);
-    }
-
-    function getBinarySync(file) {
-      if (file == wasmBinaryFile && wasmBinary) {
-        return new Uint8Array(wasmBinary);
-      }
-      if (readBinary) {
-        return readBinary(file);
-      }
-      throw 'both async and sync fetching of the wasm failed';
-    }
-
-    function getBinaryPromise(binaryFile) {
-      if (!wasmBinary && (ENVIRONMENT_IS_WEB || ENVIRONMENT_IS_WORKER)) {
-        if (typeof fetch == 'function') {
-          return fetch(binaryFile, {
-            credentials: 'same-origin',
-          })
-            .then((response) => {
-              if (!response['ok']) {
-                throw "failed to load wasm binary file at '" + binaryFile + "'";
-              }
-              return response['arrayBuffer']();
-            })
-            .catch(() => getBinarySync(binaryFile));
-        }
-      }
-      return Promise.resolve().then(() => getBinarySync(binaryFile));
-    }
-
-    function instantiateArrayBuffer(binaryFile, imports, receiver) {
-      return getBinaryPromise(binaryFile)
-        .then((binary) => WebAssembly.instantiate(binary, imports))
-        .then((instance) => instance)
-        .then(receiver, (reason) => {
-          err(`failed to asynchronously prepare wasm: ${reason}`);
-          abort(reason);
-        });
-    }
-
-    function instantiateAsync(binary, binaryFile, imports, callback) {
-      if (
-        !binary &&
-        typeof WebAssembly.instantiateStreaming == 'function' &&
-        !isDataURI(binaryFile) &&
-        typeof fetch == 'function'
-      ) {
-        return fetch(binaryFile, {
-          credentials: 'same-origin',
-        }).then((response) => {
-          /** @suppress {checkTypes} */ var result = WebAssembly.instantiateStreaming(response, imports);
-          return result.then(callback, function (reason) {
-            err(`wasm streaming compile failed: ${reason}`);
-            err('falling back to ArrayBuffer instantiation');
-            return instantiateArrayBuffer(binaryFile, imports, callback);
-          });
-        });
-      }
-      return instantiateArrayBuffer(binaryFile, imports, callback);
-    }
-
-    function createWasm() {
-      var info = {
-        a: wasmImports,
-      };
-      /** @param {WebAssembly.Module=} module*/ function receiveInstance(instance, module) {
-        wasmExports = instance.exports;
-        wasmMemory = wasmExports['sa'];
-        updateMemoryViews();
-        wasmTable = wasmExports['va'];
-        addOnInit(wasmExports['ta']);
-        removeRunDependency('wasm-instantiate');
-        return wasmExports;
-      }
-      addRunDependency('wasm-instantiate');
-      function receiveInstantiationResult(result) {
-        receiveInstance(result['instance']);
-      }
-      if (Module['instantiateWasm']) {
-        try {
-          return Module['instantiateWasm'](info, receiveInstance);
-        } catch (e) {
-          err(`Module.instantiateWasm callback failed with error: ${e}`);
-          readyPromiseReject(e);
-        }
-      }
-      instantiateAsync(wasmBinary, wasmBinaryFile, info, receiveInstantiationResult).catch(readyPromiseReject);
-      return {};
-    }
-
-    var callRuntimeCallbacks = (callbacks) => {
-      while (callbacks.length > 0) {
-        callbacks.shift()(Module);
-      }
-    };
-
-    var noExitRuntime = Module['noExitRuntime'] || true;
-
-    var UTF8Decoder = typeof TextDecoder != 'undefined' ? new TextDecoder('utf8') : undefined;
-
-    /**
-     * Given a pointer 'idx' to a null-terminated UTF8-encoded string in the given
-     * array that contains uint8 values, returns a copy of that string as a
-     * Javascript String object.
-     * heapOrArray is either a regular array, or a JavaScript typed array view.
-     * @param {number} idx
-     * @param {number=} maxBytesToRead
-     * @return {string}
-     */ var UTF8ArrayToString = (heapOrArray, idx, maxBytesToRead) => {
-      var endIdx = idx + maxBytesToRead;
-      var endPtr = idx;
-      while (heapOrArray[endPtr] && !(endPtr >= endIdx)) ++endPtr;
-      if (endPtr - idx > 16 && heapOrArray.buffer && UTF8Decoder) {
-        return UTF8Decoder.decode(heapOrArray.subarray(idx, endPtr));
-      }
-      var str = '';
-      while (idx < endPtr) {
-        var u0 = heapOrArray[idx++];
-        if (!(u0 & 128)) {
-          str += String.fromCharCode(u0);
-          continue;
-        }
-        var u1 = heapOrArray[idx++] & 63;
-        if ((u0 & 224) == 192) {
-          str += String.fromCharCode(((u0 & 31) << 6) | u1);
-          continue;
-        }
-        var u2 = heapOrArray[idx++] & 63;
-        if ((u0 & 240) == 224) {
-          u0 = ((u0 & 15) << 12) | (u1 << 6) | u2;
-        } else {
-          u0 = ((u0 & 7) << 18) | (u1 << 12) | (u2 << 6) | (heapOrArray[idx++] & 63);
-        }
-        if (u0 < 65536) {
-          str += String.fromCharCode(u0);
-        } else {
-          var ch = u0 - 65536;
-          str += String.fromCharCode(55296 | (ch >> 10), 56320 | (ch & 1023));
-        }
-      }
-      return str;
-    };
-
-    /**
-     * Given a pointer 'ptr' to a null-terminated UTF8-encoded string in the
-     * emscripten HEAP, returns a copy of that string as a Javascript String object.
-     *
-     * @param {number} ptr
-     * @param {number=} maxBytesToRead - An optional length that specifies the
-     *   maximum number of bytes to read. You can omit this parameter to scan the
-     *   string until the first 0 byte. If maxBytesToRead is passed, and the string
-     *   at [ptr, ptr+maxBytesToReadr[ contains a null byte in the middle, then the
-     *   string will cut short at that byte index (i.e. maxBytesToRead will not
-     *   produce a string of exact length [ptr, ptr+maxBytesToRead[) N.B. mixing
-     *   frequent uses of UTF8ToString() with and without maxBytesToRead may throw
-     *   JS JIT optimizations off, so it is worth to consider consistently using one
-     * @return {string}
-     */ var UTF8ToString = (ptr, maxBytesToRead) => (ptr ? UTF8ArrayToString(HEAPU8, ptr, maxBytesToRead) : '');
-
-    var ___assert_fail = (condition, filename, line, func) => {
-      abort(
-        `Assertion failed: ${UTF8ToString(condition)}, at: ` +
-          [
-            filename ? UTF8ToString(filename) : 'unknown filename',
-            line,
-            func ? UTF8ToString(func) : 'unknown function',
-          ],
-      );
-    };
-
-    var exceptionCaught = [];
-
-    var uncaughtExceptionCount = 0;
-
-    var ___cxa_begin_catch = (ptr) => {
-      var info = new ExceptionInfo(ptr);
-      if (!info.get_caught()) {
-        info.set_caught(true);
-        uncaughtExceptionCount--;
-      }
-      info.set_rethrown(false);
-      exceptionCaught.push(info);
-      ___cxa_increment_exception_refcount(info.excPtr);
-      return info.get_exception_ptr();
-    };
-
-    var exceptionLast = 0;
-
-    var ___cxa_end_catch = () => {
-      _setThrew(0, 0);
-      var info = exceptionCaught.pop();
-      ___cxa_decrement_exception_refcount(info.excPtr);
-      exceptionLast = 0;
-    };
-
-    /** @constructor */ function ExceptionInfo(excPtr) {
-      this.excPtr = excPtr;
-      this.ptr = excPtr - 24;
-      this.set_type = function (type) {
-        HEAPU32[(this.ptr + 4) >> 2] = type;
-      };
-      this.get_type = function () {
-        return HEAPU32[(this.ptr + 4) >> 2];
-      };
-      this.set_destructor = function (destructor) {
-        HEAPU32[(this.ptr + 8) >> 2] = destructor;
-      };
-      this.get_destructor = function () {
-        return HEAPU32[(this.ptr + 8) >> 2];
-      };
-      this.set_caught = function (caught) {
-        caught = caught ? 1 : 0;
-        HEAP8[(this.ptr + 12) >> 0] = caught;
-      };
-      this.get_caught = function () {
-        return HEAP8[(this.ptr + 12) >> 0] != 0;
-      };
-      this.set_rethrown = function (rethrown) {
-        rethrown = rethrown ? 1 : 0;
-        HEAP8[(this.ptr + 13) >> 0] = rethrown;
-      };
-      this.get_rethrown = function () {
-        return HEAP8[(this.ptr + 13) >> 0] != 0;
-      };
-      this.init = function (type, destructor) {
-        this.set_adjusted_ptr(0);
-        this.set_type(type);
-        this.set_destructor(destructor);
-      };
-      this.set_adjusted_ptr = function (adjustedPtr) {
-        HEAPU32[(this.ptr + 16) >> 2] = adjustedPtr;
-      };
-      this.get_adjusted_ptr = function () {
-        return HEAPU32[(this.ptr + 16) >> 2];
-      };
-      this.get_exception_ptr = function () {
-        var isPointer = ___cxa_is_pointer_type(this.get_type());
-        if (isPointer) {
-          return HEAPU32[this.excPtr >> 2];
-        }
-        var adjusted = this.get_adjusted_ptr();
-        if (adjusted !== 0) return adjusted;
-        return this.excPtr;
-      };
-    }
-
-    var ___resumeException = (ptr) => {
-      if (!exceptionLast) {
-        exceptionLast = ptr;
-      }
-      throw exceptionLast;
-    };
-
-    var findMatchingCatch = (args) => {
-      var thrown = exceptionLast;
-      if (!thrown) {
-        setTempRet0(0);
-        return 0;
-      }
-      var info = new ExceptionInfo(thrown);
-      info.set_adjusted_ptr(thrown);
-      var thrownType = info.get_type();
-      if (!thrownType) {
-        setTempRet0(0);
-        return thrown;
-      }
-      for (var arg in args) {
-        var caughtType = args[arg];
-        if (caughtType === 0 || caughtType === thrownType) {
-          break;
-        }
-        var adjusted_ptr_addr = info.ptr + 16;
-        if (___cxa_can_catch(caughtType, thrownType, adjusted_ptr_addr)) {
-          setTempRet0(caughtType);
-          return thrown;
-        }
-      }
-      setTempRet0(thrownType);
-      return thrown;
-    };
-
-    var ___cxa_find_matching_catch_2 = () => findMatchingCatch([]);
-
-    var ___cxa_find_matching_catch_4 = (arg0, arg1) => findMatchingCatch([arg0, arg1]);
-
-    var ___cxa_throw = (ptr, type, destructor) => {
-      var info = new ExceptionInfo(ptr);
-      info.init(type, destructor);
-      exceptionLast = ptr;
-      uncaughtExceptionCount++;
-      throw exceptionLast;
-    };
-
-    var SYSCALLS = {
-      varargs: undefined,
-      get() {
-        var ret = HEAP32[+SYSCALLS.varargs >> 2];
-        SYSCALLS.varargs += 4;
-        return ret;
-      },
-      getp() {
-        return SYSCALLS.get();
-      },
-      getStr(ptr) {
-        var ret = UTF8ToString(ptr);
-        return ret;
-      },
-    };
-
-    function ___syscall_fcntl64(fd, cmd, varargs) {
-      SYSCALLS.varargs = varargs;
-      return 0;
-    }
-
-    var lengthBytesUTF8 = (str) => {
-      var len = 0;
-      for (var i = 0; i < str.length; ++i) {
-        var c = str.charCodeAt(i);
-        if (c <= 127) {
-          len++;
-        } else if (c <= 2047) {
-          len += 2;
-        } else if (c >= 55296 && c <= 57343) {
-          len += 4;
-          ++i;
-        } else {
-          len += 3;
-        }
-      }
-      return len;
-    };
-
-    var stringToUTF8Array = (str, heap, outIdx, maxBytesToWrite) => {
-      if (!(maxBytesToWrite > 0)) return 0;
-      var startIdx = outIdx;
-      var endIdx = outIdx + maxBytesToWrite - 1;
-      for (var i = 0; i < str.length; ++i) {
-        var u = str.charCodeAt(i);
-        if (u >= 55296 && u <= 57343) {
-          var u1 = str.charCodeAt(++i);
-          u = (65536 + ((u & 1023) << 10)) | (u1 & 1023);
-        }
-        if (u <= 127) {
-          if (outIdx >= endIdx) break;
-          heap[outIdx++] = u;
-        } else if (u <= 2047) {
-          if (outIdx + 1 >= endIdx) break;
-          heap[outIdx++] = 192 | (u >> 6);
-          heap[outIdx++] = 128 | (u & 63);
-        } else if (u <= 65535) {
-          if (outIdx + 2 >= endIdx) break;
-          heap[outIdx++] = 224 | (u >> 12);
-          heap[outIdx++] = 128 | ((u >> 6) & 63);
-          heap[outIdx++] = 128 | (u & 63);
-        } else {
-          if (outIdx + 3 >= endIdx) break;
-          heap[outIdx++] = 240 | (u >> 18);
-          heap[outIdx++] = 128 | ((u >> 12) & 63);
-          heap[outIdx++] = 128 | ((u >> 6) & 63);
-          heap[outIdx++] = 128 | (u & 63);
-        }
-      }
-      heap[outIdx] = 0;
-      return outIdx - startIdx;
-    };
-
-    var stringToUTF8 = (str, outPtr, maxBytesToWrite) => stringToUTF8Array(str, HEAPU8, outPtr, maxBytesToWrite);
-
-    var ___syscall_getcwd = (buf, size) => {};
-
-    function ___syscall_ioctl(fd, op, varargs) {
-      SYSCALLS.varargs = varargs;
-      return 0;
-    }
-
-    function ___syscall_openat(dirfd, path, flags, varargs) {
-      SYSCALLS.varargs = varargs;
-    }
-
-    var structRegistrations = {};
-
-    var runDestructors = (destructors) => {
-      while (destructors.length) {
-        var ptr = destructors.pop();
-        var del = destructors.pop();
-        del(ptr);
-      }
-    };
-
-    /** @suppress {globalThis} */ function simpleReadValueFromPointer(pointer) {
-      return this['fromWireType'](HEAP32[pointer >> 2]);
-    }
-
-    var awaitingDependencies = {};
-
-    var registeredTypes = {};
-
-    var typeDependencies = {};
-
-    var InternalError;
-
-    var throwInternalError = (message) => {
-      throw new InternalError(message);
-    };
-
-    var whenDependentTypesAreResolved = (myTypes, dependentTypes, getTypeConverters) => {
-      myTypes.forEach(function (type) {
-        typeDependencies[type] = dependentTypes;
-      });
-      function onComplete(typeConverters) {
-        var myTypeConverters = getTypeConverters(typeConverters);
-        if (myTypeConverters.length !== myTypes.length) {
-          throwInternalError('Mismatched type converter count');
-        }
-        for (var i = 0; i < myTypes.length; ++i) {
-          registerType(myTypes[i], myTypeConverters[i]);
-        }
-      }
-      var typeConverters = new Array(dependentTypes.length);
-      var unregisteredTypes = [];
-      var registered = 0;
-      dependentTypes.forEach((dt, i) => {
-        if (registeredTypes.hasOwnProperty(dt)) {
-          typeConverters[i] = registeredTypes[dt];
-        } else {
-          unregisteredTypes.push(dt);
-          if (!awaitingDependencies.hasOwnProperty(dt)) {
-            awaitingDependencies[dt] = [];
-          }
-          awaitingDependencies[dt].push(() => {
-            typeConverters[i] = registeredTypes[dt];
-            ++registered;
-            if (registered === unregisteredTypes.length) {
-              onComplete(typeConverters);
-            }
-          });
-        }
-      });
-      if (0 === unregisteredTypes.length) {
-        onComplete(typeConverters);
-      }
-    };
-
-    var __embind_finalize_value_object = (structType) => {
-      var reg = structRegistrations[structType];
-      delete structRegistrations[structType];
-      var rawConstructor = reg.rawConstructor;
-      var rawDestructor = reg.rawDestructor;
-      var fieldRecords = reg.fields;
-      var fieldTypes = fieldRecords
-        .map((field) => field.getterReturnType)
-        .concat(fieldRecords.map((field) => field.setterArgumentType));
-      whenDependentTypesAreResolved([structType], fieldTypes, (fieldTypes) => {
-        var fields = {};
-        fieldRecords.forEach((field, i) => {
-          var fieldName = field.fieldName;
-          var getterReturnType = fieldTypes[i];
-          var getter = field.getter;
-          var getterContext = field.getterContext;
-          var setterArgumentType = fieldTypes[i + fieldRecords.length];
-          var setter = field.setter;
-          var setterContext = field.setterContext;
-          fields[fieldName] = {
-            read: (ptr) => getterReturnType['fromWireType'](getter(getterContext, ptr)),
-            write: (ptr, o) => {
-              var destructors = [];
-              setter(setterContext, ptr, setterArgumentType['toWireType'](destructors, o));
-              runDestructors(destructors);
-            },
-          };
-        });
-        return [
-          {
-            name: reg.name,
-            fromWireType: (ptr) => {
-              var rv = {};
-              for (var i in fields) {
-                rv[i] = fields[i].read(ptr);
-              }
-              rawDestructor(ptr);
-              return rv;
-            },
-            toWireType: (destructors, o) => {
-              for (var fieldName in fields) {
-                if (!(fieldName in o)) {
-                  throw new TypeError(`Missing field: "${fieldName}"`);
-                }
-              }
-              var ptr = rawConstructor();
-              for (fieldName in fields) {
-                fields[fieldName].write(ptr, o[fieldName]);
-              }
-              if (destructors !== null) {
-                destructors.push(rawDestructor, ptr);
-              }
-              return ptr;
-            },
-            argPackAdvance: GenericWireTypeSize,
-            readValueFromPointer: simpleReadValueFromPointer,
-            destructorFunction: rawDestructor,
-          },
-        ];
-      });
-    };
-
-    var __embind_register_bigint = (primitiveType, name, size, minRange, maxRange) => {};
-
-    var embind_init_charCodes = () => {
-      var codes = new Array(256);
-      for (var i = 0; i < 256; ++i) {
-        codes[i] = String.fromCharCode(i);
-      }
-      embind_charCodes = codes;
-    };
-
-    var embind_charCodes;
-
-    var readLatin1String = (ptr) => {
-      var ret = '';
-      var c = ptr;
-      while (HEAPU8[c]) {
-        ret += embind_charCodes[HEAPU8[c++]];
-      }
-      return ret;
-    };
-
-    var BindingError;
-
-    var throwBindingError = (message) => {
-      throw new BindingError(message);
-    };
-
-    /** @param {Object=} options */ function sharedRegisterType(rawType, registeredInstance, options = {}) {
-      var name = registeredInstance.name;
-      if (!rawType) {
-        throwBindingError(`type "${name}" must have a positive integer typeid pointer`);
-      }
-      if (registeredTypes.hasOwnProperty(rawType)) {
-        if (options.ignoreDuplicateRegistrations) {
-          return;
-        } else {
-          throwBindingError(`Cannot register type '${name}' twice`);
-        }
-      }
-      registeredTypes[rawType] = registeredInstance;
-      delete typeDependencies[rawType];
-      if (awaitingDependencies.hasOwnProperty(rawType)) {
-        var callbacks = awaitingDependencies[rawType];
-        delete awaitingDependencies[rawType];
-        callbacks.forEach((cb) => cb());
-      }
-    }
-
-    /** @param {Object=} options */ function registerType(rawType, registeredInstance, options = {}) {
-      if (!('argPackAdvance' in registeredInstance)) {
-        throw new TypeError('registerType registeredInstance requires argPackAdvance');
-      }
-      return sharedRegisterType(rawType, registeredInstance, options);
-    }
-
-    var GenericWireTypeSize = 8;
-
-    /** @suppress {globalThis} */ var __embind_register_bool = (rawType, name, trueValue, falseValue) => {
-      name = readLatin1String(name);
-      registerType(rawType, {
-        name: name,
-        fromWireType: function (wt) {
-          return !!wt;
-        },
-        toWireType: function (destructors, o) {
-          return o ? trueValue : falseValue;
-        },
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: function (pointer) {
-          return this['fromWireType'](HEAPU8[pointer]);
-        },
-        destructorFunction: null,
-      });
-    };
-
-    var shallowCopyInternalPointer = (o) => ({
-      count: o.count,
-      deleteScheduled: o.deleteScheduled,
-      preservePointerOnDelete: o.preservePointerOnDelete,
-      ptr: o.ptr,
-      ptrType: o.ptrType,
-      smartPtr: o.smartPtr,
-      smartPtrType: o.smartPtrType,
-    });
-
-    var throwInstanceAlreadyDeleted = (obj) => {
-      function getInstanceTypeName(handle) {
-        return handle.$$.ptrType.registeredClass.name;
-      }
-      throwBindingError(getInstanceTypeName(obj) + ' instance already deleted');
-    };
-
-    var finalizationRegistry = false;
-
-    var detachFinalizer = (handle) => {};
-
-    var runDestructor = ($$) => {
-      if ($$.smartPtr) {
-        $$.smartPtrType.rawDestructor($$.smartPtr);
-      } else {
-        $$.ptrType.registeredClass.rawDestructor($$.ptr);
-      }
-    };
-
-    var releaseClassHandle = ($$) => {
-      $$.count.value -= 1;
-      var toDelete = 0 === $$.count.value;
-      if (toDelete) {
-        runDestructor($$);
-      }
-    };
-
-    var downcastPointer = (ptr, ptrClass, desiredClass) => {
-      if (ptrClass === desiredClass) {
-        return ptr;
-      }
-      if (undefined === desiredClass.baseClass) {
-        return null;
-      }
-      var rv = downcastPointer(ptr, ptrClass, desiredClass.baseClass);
-      if (rv === null) {
-        return null;
-      }
-      return desiredClass.downcast(rv);
-    };
-
-    var registeredPointers = {};
-
-    var getInheritedInstanceCount = () => Object.keys(registeredInstances).length;
-
-    var getLiveInheritedInstances = () => {
-      var rv = [];
-      for (var k in registeredInstances) {
-        if (registeredInstances.hasOwnProperty(k)) {
-          rv.push(registeredInstances[k]);
-        }
-      }
-      return rv;
-    };
-
-    var deletionQueue = [];
-
-    var flushPendingDeletes = () => {
-      while (deletionQueue.length) {
-        var obj = deletionQueue.pop();
-        obj.$$.deleteScheduled = false;
-        obj['delete']();
-      }
-    };
-
-    var delayFunction;
-
-    var setDelayFunction = (fn) => {
-      delayFunction = fn;
-      if (deletionQueue.length && delayFunction) {
-        delayFunction(flushPendingDeletes);
-      }
-    };
-
-    var init_embind = () => {
-      Module['getInheritedInstanceCount'] = getInheritedInstanceCount;
-      Module['getLiveInheritedInstances'] = getLiveInheritedInstances;
-      Module['flushPendingDeletes'] = flushPendingDeletes;
-      Module['setDelayFunction'] = setDelayFunction;
-    };
-
-    var registeredInstances = {};
-
-    var getBasestPointer = (class_, ptr) => {
-      if (ptr === undefined) {
-        throwBindingError('ptr should not be undefined');
-      }
-      while (class_.baseClass) {
-        ptr = class_.upcast(ptr);
-        class_ = class_.baseClass;
-      }
-      return ptr;
-    };
-
-    var getInheritedInstance = (class_, ptr) => {
-      ptr = getBasestPointer(class_, ptr);
-      return registeredInstances[ptr];
-    };
-
-    var makeClassHandle = (prototype, record) => {
-      if (!record.ptrType || !record.ptr) {
-        throwInternalError('makeClassHandle requires ptr and ptrType');
-      }
-      var hasSmartPtrType = !!record.smartPtrType;
-      var hasSmartPtr = !!record.smartPtr;
-      if (hasSmartPtrType !== hasSmartPtr) {
-        throwInternalError('Both smartPtrType and smartPtr must be specified');
-      }
-      record.count = {
-        value: 1,
-      };
-      return attachFinalizer(
-        Object.create(prototype, {
-          $$: {
-            value: record,
-            writable: true,
-          },
-        }),
-      );
-    };
-
-    /** @suppress {globalThis} */ function RegisteredPointer_fromWireType(ptr) {
-      var rawPointer = this.getPointee(ptr);
-      if (!rawPointer) {
-        this.destructor(ptr);
-        return null;
-      }
-      var registeredInstance = getInheritedInstance(this.registeredClass, rawPointer);
-      if (undefined !== registeredInstance) {
-        if (0 === registeredInstance.$$.count.value) {
-          registeredInstance.$$.ptr = rawPointer;
-          registeredInstance.$$.smartPtr = ptr;
-          return registeredInstance['clone']();
-        } else {
-          var rv = registeredInstance['clone']();
-          this.destructor(ptr);
-          return rv;
-        }
-      }
-      function makeDefaultHandle() {
-        if (this.isSmartPointer) {
-          return makeClassHandle(this.registeredClass.instancePrototype, {
-            ptrType: this.pointeeType,
-            ptr: rawPointer,
-            smartPtrType: this,
-            smartPtr: ptr,
-          });
-        } else {
-          return makeClassHandle(this.registeredClass.instancePrototype, {
-            ptrType: this,
-            ptr: ptr,
-          });
-        }
-      }
-      var actualType = this.registeredClass.getActualType(rawPointer);
-      var registeredPointerRecord = registeredPointers[actualType];
-      if (!registeredPointerRecord) {
-        return makeDefaultHandle.call(this);
-      }
-      var toType;
-      if (this.isConst) {
-        toType = registeredPointerRecord.constPointerType;
-      } else {
-        toType = registeredPointerRecord.pointerType;
-      }
-      var dp = downcastPointer(rawPointer, this.registeredClass, toType.registeredClass);
-      if (dp === null) {
-        return makeDefaultHandle.call(this);
-      }
-      if (this.isSmartPointer) {
-        return makeClassHandle(toType.registeredClass.instancePrototype, {
-          ptrType: toType,
-          ptr: dp,
-          smartPtrType: this,
-          smartPtr: ptr,
-        });
-      } else {
-        return makeClassHandle(toType.registeredClass.instancePrototype, {
-          ptrType: toType,
-          ptr: dp,
-        });
-      }
-    }
-
-    var attachFinalizer = (handle) => {
-      if ('undefined' === typeof FinalizationRegistry) {
-        attachFinalizer = (handle) => handle;
-        return handle;
-      }
-      finalizationRegistry = new FinalizationRegistry((info) => {
-        releaseClassHandle(info.$$);
-      });
-      attachFinalizer = (handle) => {
-        var $$ = handle.$$;
-        var hasSmartPtr = !!$$.smartPtr;
-        if (hasSmartPtr) {
-          var info = {
-            $$: $$,
-          };
-          finalizationRegistry.register(handle, info, handle);
-        }
-        return handle;
-      };
-      detachFinalizer = (handle) => finalizationRegistry.unregister(handle);
-      return attachFinalizer(handle);
-    };
-
-    var init_ClassHandle = () => {
-      Object.assign(ClassHandle.prototype, {
-        isAliasOf(other) {
-          if (!(this instanceof ClassHandle)) {
-            return false;
-          }
-          if (!(other instanceof ClassHandle)) {
-            return false;
-          }
-          var leftClass = this.$$.ptrType.registeredClass;
-          var left = this.$$.ptr;
-          other.$$ = /** @type {Object} */ (other.$$);
-          var rightClass = other.$$.ptrType.registeredClass;
-          var right = other.$$.ptr;
-          while (leftClass.baseClass) {
-            left = leftClass.upcast(left);
-            leftClass = leftClass.baseClass;
-          }
-          while (rightClass.baseClass) {
-            right = rightClass.upcast(right);
-            rightClass = rightClass.baseClass;
-          }
-          return leftClass === rightClass && left === right;
-        },
-        clone() {
-          if (!this.$$.ptr) {
-            throwInstanceAlreadyDeleted(this);
-          }
-          if (this.$$.preservePointerOnDelete) {
-            this.$$.count.value += 1;
-            return this;
-          } else {
-            var clone = attachFinalizer(
-              Object.create(Object.getPrototypeOf(this), {
-                $$: {
-                  value: shallowCopyInternalPointer(this.$$),
-                },
-              }),
-            );
-            clone.$$.count.value += 1;
-            clone.$$.deleteScheduled = false;
-            return clone;
-          }
-        },
-        delete() {
-          if (!this.$$.ptr) {
-            throwInstanceAlreadyDeleted(this);
-          }
-          if (this.$$.deleteScheduled && !this.$$.preservePointerOnDelete) {
-            throwBindingError('Object already scheduled for deletion');
-          }
-          detachFinalizer(this);
-          releaseClassHandle(this.$$);
-          if (!this.$$.preservePointerOnDelete) {
-            this.$$.smartPtr = undefined;
-            this.$$.ptr = undefined;
-          }
-        },
-        isDeleted() {
-          return !this.$$.ptr;
-        },
-        deleteLater() {
-          if (!this.$$.ptr) {
-            throwInstanceAlreadyDeleted(this);
-          }
-          if (this.$$.deleteScheduled && !this.$$.preservePointerOnDelete) {
-            throwBindingError('Object already scheduled for deletion');
-          }
-          deletionQueue.push(this);
-          if (deletionQueue.length === 1 && delayFunction) {
-            delayFunction(flushPendingDeletes);
-          }
-          this.$$.deleteScheduled = true;
-          return this;
-        },
-      });
-    };
-
-    /** @constructor */ function ClassHandle() {}
-
-    var createNamedFunction = (name, body) =>
-      Object.defineProperty(body, 'name', {
-        value: name,
-      });
-
-    var ensureOverloadTable = (proto, methodName, humanName) => {
-      if (undefined === proto[methodName].overloadTable) {
-        var prevFunc = proto[methodName];
-        proto[methodName] = function () {
-          if (!proto[methodName].overloadTable.hasOwnProperty(arguments.length)) {
-            throwBindingError(
-              `Function '${humanName}' called with an invalid number of arguments (${arguments.length}) - expects one of (${proto[methodName].overloadTable})!`,
-            );
-          }
-          return proto[methodName].overloadTable[arguments.length].apply(this, arguments);
-        };
-        proto[methodName].overloadTable = [];
-        proto[methodName].overloadTable[prevFunc.argCount] = prevFunc;
-      }
-    };
-
-    /** @param {number=} numArguments */ var exposePublicSymbol = (name, value, numArguments) => {
-      if (Module.hasOwnProperty(name)) {
-        if (
-          undefined === numArguments ||
-          (undefined !== Module[name].overloadTable && undefined !== Module[name].overloadTable[numArguments])
-        ) {
-          throwBindingError(`Cannot register public name '${name}' twice`);
-        }
-        ensureOverloadTable(Module, name, name);
-        if (Module.hasOwnProperty(numArguments)) {
-          throwBindingError(
-            `Cannot register multiple overloads of a function with the same number of arguments (${numArguments})!`,
-          );
-        }
-        Module[name].overloadTable[numArguments] = value;
-      } else {
-        Module[name] = value;
-        if (undefined !== numArguments) {
-          Module[name].numArguments = numArguments;
-        }
-      }
-    };
-
-    var char_0 = 48;
-
-    var char_9 = 57;
-
-    var makeLegalFunctionName = (name) => {
-      if (undefined === name) {
-        return '_unknown';
-      }
-      name = name.replace(/[^a-zA-Z0-9_]/g, '$');
-      var f = name.charCodeAt(0);
-      if (f >= char_0 && f <= char_9) {
-        return `_${name}`;
-      }
-      return name;
-    };
-
-    /** @constructor */ function RegisteredClass(
-      name,
-      constructor,
-      instancePrototype,
-      rawDestructor,
-      baseClass,
-      getActualType,
-      upcast,
-      downcast,
-    ) {
-      this.name = name;
-      this.constructor = constructor;
-      this.instancePrototype = instancePrototype;
-      this.rawDestructor = rawDestructor;
-      this.baseClass = baseClass;
-      this.getActualType = getActualType;
-      this.upcast = upcast;
-      this.downcast = downcast;
-      this.pureVirtualFunctions = [];
-    }
-
-    var upcastPointer = (ptr, ptrClass, desiredClass) => {
-      while (ptrClass !== desiredClass) {
-        if (!ptrClass.upcast) {
-          throwBindingError(`Expected null or instance of ${desiredClass.name}, got an instance of ${ptrClass.name}`);
-        }
-        ptr = ptrClass.upcast(ptr);
-        ptrClass = ptrClass.baseClass;
-      }
-      return ptr;
-    };
-
-    /** @suppress {globalThis} */ function constNoSmartPtrRawPointerToWireType(destructors, handle) {
-      if (handle === null) {
-        if (this.isReference) {
-          throwBindingError(`null is not a valid ${this.name}`);
-        }
-        return 0;
-      }
-      if (!handle.$$) {
-        throwBindingError(`Cannot pass "${embindRepr(handle)}" as a ${this.name}`);
-      }
-      if (!handle.$$.ptr) {
-        throwBindingError(`Cannot pass deleted object as a pointer of type ${this.name}`);
-      }
-      var handleClass = handle.$$.ptrType.registeredClass;
-      var ptr = upcastPointer(handle.$$.ptr, handleClass, this.registeredClass);
-      return ptr;
-    }
-
-    /** @suppress {globalThis} */ function genericPointerToWireType(destructors, handle) {
-      var ptr;
-      if (handle === null) {
-        if (this.isReference) {
-          throwBindingError(`null is not a valid ${this.name}`);
-        }
-        if (this.isSmartPointer) {
-          ptr = this.rawConstructor();
-          if (destructors !== null) {
-            destructors.push(this.rawDestructor, ptr);
-          }
-          return ptr;
-        } else {
-          return 0;
-        }
-      }
-      if (!handle || !handle.$$) {
-        throwBindingError(`Cannot pass "${embindRepr(handle)}" as a ${this.name}`);
-      }
-      if (!handle.$$.ptr) {
-        throwBindingError(`Cannot pass deleted object as a pointer of type ${this.name}`);
-      }
-      if (!this.isConst && handle.$$.ptrType.isConst) {
-        throwBindingError(
-          `Cannot convert argument of type ${
-            handle.$$.smartPtrType ? handle.$$.smartPtrType.name : handle.$$.ptrType.name
-          } to parameter type ${this.name}`,
-        );
-      }
-      var handleClass = handle.$$.ptrType.registeredClass;
-      ptr = upcastPointer(handle.$$.ptr, handleClass, this.registeredClass);
-      if (this.isSmartPointer) {
-        if (undefined === handle.$$.smartPtr) {
-          throwBindingError('Passing raw pointer to smart pointer is illegal');
-        }
-        switch (this.sharingPolicy) {
-          case 0:
-            if (handle.$$.smartPtrType === this) {
-              ptr = handle.$$.smartPtr;
-            } else {
-              throwBindingError(
-                `Cannot convert argument of type ${
-                  handle.$$.smartPtrType ? handle.$$.smartPtrType.name : handle.$$.ptrType.name
-                } to parameter type ${this.name}`,
-              );
-            }
-            break;
-
-          case 1:
-            ptr = handle.$$.smartPtr;
-            break;
-
-          case 2:
-            if (handle.$$.smartPtrType === this) {
-              ptr = handle.$$.smartPtr;
-            } else {
-              var clonedHandle = handle['clone']();
-              ptr = this.rawShare(
-                ptr,
-                Emval.toHandle(() => clonedHandle['delete']()),
-              );
-              if (destructors !== null) {
-                destructors.push(this.rawDestructor, ptr);
-              }
-            }
-            break;
-
-          default:
-            throwBindingError('Unsupporting sharing policy');
-        }
-      }
-      return ptr;
-    }
-
-    /** @suppress {globalThis} */ function nonConstNoSmartPtrRawPointerToWireType(destructors, handle) {
-      if (handle === null) {
-        if (this.isReference) {
-          throwBindingError(`null is not a valid ${this.name}`);
-        }
-        return 0;
-      }
-      if (!handle.$$) {
-        throwBindingError(`Cannot pass "${embindRepr(handle)}" as a ${this.name}`);
-      }
-      if (!handle.$$.ptr) {
-        throwBindingError(`Cannot pass deleted object as a pointer of type ${this.name}`);
-      }
-      if (handle.$$.ptrType.isConst) {
-        throwBindingError(`Cannot convert argument of type ${handle.$$.ptrType.name} to parameter type ${this.name}`);
-      }
-      var handleClass = handle.$$.ptrType.registeredClass;
-      var ptr = upcastPointer(handle.$$.ptr, handleClass, this.registeredClass);
-      return ptr;
-    }
-
-    /** @suppress {globalThis} */ function readPointer(pointer) {
-      return this['fromWireType'](HEAPU32[pointer >> 2]);
-    }
-
-    var init_RegisteredPointer = () => {
-      Object.assign(RegisteredPointer.prototype, {
-        getPointee(ptr) {
-          if (this.rawGetPointee) {
-            ptr = this.rawGetPointee(ptr);
-          }
-          return ptr;
-        },
-        destructor(ptr) {
-          this.rawDestructor?.(ptr);
-        },
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: readPointer,
-        deleteObject(handle) {
-          if (handle !== null) {
-            handle['delete']();
-          }
-        },
-        fromWireType: RegisteredPointer_fromWireType,
-      });
-    };
-
-    /** @constructor
-      @param {*=} pointeeType,
-      @param {*=} sharingPolicy,
-      @param {*=} rawGetPointee,
-      @param {*=} rawConstructor,
-      @param {*=} rawShare,
-      @param {*=} rawDestructor,
-       */ function RegisteredPointer(
-      name,
-      registeredClass,
-      isReference,
-      isConst,
-      isSmartPointer,
-      pointeeType,
-      sharingPolicy,
-      rawGetPointee,
-      rawConstructor,
-      rawShare,
-      rawDestructor,
-    ) {
-      this.name = name;
-      this.registeredClass = registeredClass;
-      this.isReference = isReference;
-      this.isConst = isConst;
-      this.isSmartPointer = isSmartPointer;
-      this.pointeeType = pointeeType;
-      this.sharingPolicy = sharingPolicy;
-      this.rawGetPointee = rawGetPointee;
-      this.rawConstructor = rawConstructor;
-      this.rawShare = rawShare;
-      this.rawDestructor = rawDestructor;
-      if (!isSmartPointer && registeredClass.baseClass === undefined) {
-        if (isConst) {
-          this['toWireType'] = constNoSmartPtrRawPointerToWireType;
-          this.destructorFunction = null;
-        } else {
-          this['toWireType'] = nonConstNoSmartPtrRawPointerToWireType;
-          this.destructorFunction = null;
-        }
-      } else {
-        this['toWireType'] = genericPointerToWireType;
-      }
-    }
-
-    /** @param {number=} numArguments */ var replacePublicSymbol = (name, value, numArguments) => {
-      if (!Module.hasOwnProperty(name)) {
-        throwInternalError('Replacing nonexistant public symbol');
-      }
-      if (undefined !== Module[name].overloadTable && undefined !== numArguments) {
-        Module[name].overloadTable[numArguments] = value;
-      } else {
-        Module[name] = value;
-        Module[name].argCount = numArguments;
-      }
-    };
-
-    var dynCallLegacy = (sig, ptr, args) => {
-      var f = Module['dynCall_' + sig];
-      return args && args.length ? f.apply(null, [ptr].concat(args)) : f.call(null, ptr);
-    };
-
-    var wasmTable;
-
-    var getWasmTableEntry = (funcPtr) => wasmTable.get(funcPtr);
-
-    /** @param {Object=} args */ var dynCall = (sig, ptr, args) => {
-      if (sig.includes('j')) {
-        return dynCallLegacy(sig, ptr, args);
-      }
-      var rtn = getWasmTableEntry(ptr).apply(null, args);
-      return rtn;
-    };
-
-    var getDynCaller = (sig, ptr) => {
-      var argCache = [];
-      return function () {
-        argCache.length = 0;
-        Object.assign(argCache, arguments);
-        return dynCall(sig, ptr, argCache);
-      };
-    };
-
-    var embind__requireFunction = (signature, rawFunction) => {
-      signature = readLatin1String(signature);
-      function makeDynCaller() {
-        if (signature.includes('j')) {
-          return getDynCaller(signature, rawFunction);
-        }
-        return getWasmTableEntry(rawFunction);
-      }
-      var fp = makeDynCaller();
-      if (typeof fp != 'function') {
-        throwBindingError(`unknown function pointer with signature ${signature}: ${rawFunction}`);
-      }
-      return fp;
-    };
-
-    var extendError = (baseErrorType, errorName) => {
-      var errorClass = createNamedFunction(errorName, function (message) {
-        this.name = errorName;
-        this.message = message;
-        var stack = new Error(message).stack;
-        if (stack !== undefined) {
-          this.stack = this.toString() + '\n' + stack.replace(/^Error(:[^\n]*)?\n/, '');
-        }
-      });
-      errorClass.prototype = Object.create(baseErrorType.prototype);
-      errorClass.prototype.constructor = errorClass;
-      errorClass.prototype.toString = function () {
-        if (this.message === undefined) {
-          return this.name;
-        } else {
-          return `${this.name}: ${this.message}`;
-        }
-      };
-      return errorClass;
-    };
-
-    var UnboundTypeError;
-
-    var getTypeName = (type) => {
-      var ptr = ___getTypeName(type);
-      var rv = readLatin1String(ptr);
-      _free(ptr);
-      return rv;
-    };
-
-    var throwUnboundTypeError = (message, types) => {
-      var unboundTypes = [];
-      var seen = {};
-      function visit(type) {
-        if (seen[type]) {
-          return;
-        }
-        if (registeredTypes[type]) {
-          return;
-        }
-        if (typeDependencies[type]) {
-          typeDependencies[type].forEach(visit);
-          return;
-        }
-        unboundTypes.push(type);
-        seen[type] = true;
-      }
-      types.forEach(visit);
-      throw new UnboundTypeError(`${message}: ` + unboundTypes.map(getTypeName).join([', ']));
-    };
-
-    var __embind_register_class = (
-      rawType,
-      rawPointerType,
-      rawConstPointerType,
-      baseClassRawType,
-      getActualTypeSignature,
-      getActualType,
-      upcastSignature,
-      upcast,
-      downcastSignature,
-      downcast,
-      name,
-      destructorSignature,
-      rawDestructor,
-    ) => {
-      name = readLatin1String(name);
-      getActualType = embind__requireFunction(getActualTypeSignature, getActualType);
-      upcast &&= embind__requireFunction(upcastSignature, upcast);
-      downcast &&= embind__requireFunction(downcastSignature, downcast);
-      rawDestructor = embind__requireFunction(destructorSignature, rawDestructor);
-      var legalFunctionName = makeLegalFunctionName(name);
-      exposePublicSymbol(legalFunctionName, function () {
-        throwUnboundTypeError(`Cannot construct ${name} due to unbound types`, [baseClassRawType]);
-      });
-      whenDependentTypesAreResolved(
-        [rawType, rawPointerType, rawConstPointerType],
-        baseClassRawType ? [baseClassRawType] : [],
-        function (base) {
-          base = base[0];
-          var baseClass;
-          var basePrototype;
-          if (baseClassRawType) {
-            baseClass = base.registeredClass;
-            basePrototype = baseClass.instancePrototype;
-          } else {
-            basePrototype = ClassHandle.prototype;
-          }
-          var constructor = createNamedFunction(name, function () {
-            if (Object.getPrototypeOf(this) !== instancePrototype) {
-              throw new BindingError("Use 'new' to construct " + name);
-            }
-            if (undefined === registeredClass.constructor_body) {
-              throw new BindingError(name + ' has no accessible constructor');
-            }
-            var body = registeredClass.constructor_body[arguments.length];
-            if (undefined === body) {
-              throw new BindingError(
-                `Tried to invoke ctor of ${name} with invalid number of parameters (${
-                  arguments.length
-                }) - expected (${Object.keys(registeredClass.constructor_body).toString()}) parameters instead!`,
-              );
-            }
-            return body.apply(this, arguments);
-          });
-          var instancePrototype = Object.create(basePrototype, {
-            constructor: {
-              value: constructor,
-            },
-          });
-          constructor.prototype = instancePrototype;
-          var registeredClass = new RegisteredClass(
-            name,
-            constructor,
-            instancePrototype,
-            rawDestructor,
-            baseClass,
-            getActualType,
-            upcast,
-            downcast,
-          );
-          if (registeredClass.baseClass) {
-            registeredClass.baseClass.__derivedClasses ??= [];
-            registeredClass.baseClass.__derivedClasses.push(registeredClass);
-          }
-          var referenceConverter = new RegisteredPointer(name, registeredClass, true, false, false);
-          var pointerConverter = new RegisteredPointer(name + '*', registeredClass, false, false, false);
-          var constPointerConverter = new RegisteredPointer(name + ' const*', registeredClass, false, true, false);
-          registeredPointers[rawType] = {
-            pointerType: pointerConverter,
-            constPointerType: constPointerConverter,
-          };
-          replacePublicSymbol(legalFunctionName, constructor);
-          return [referenceConverter, pointerConverter, constPointerConverter];
-        },
-      );
-    };
-
-    var heap32VectorToArray = (count, firstElement) => {
-      var array = [];
-      for (var i = 0; i < count; i++) {
-        array.push(HEAPU32[(firstElement + i * 4) >> 2]);
-      }
-      return array;
-    };
-
-    function usesDestructorStack(argTypes) {
-      for (var i = 1; i < argTypes.length; ++i) {
-        if (argTypes[i] !== null && argTypes[i].destructorFunction === undefined) {
-          return true;
-        }
-      }
-      return false;
-    }
-
-    function craftInvokerFunction(
-      humanName,
-      argTypes,
-      classType,
-      cppInvokerFunc,
-      cppTargetFunc,
-      /** boolean= */ isAsync,
-    ) {
-      var argCount = argTypes.length;
-      if (argCount < 2) {
-        throwBindingError("argTypes array size mismatch! Must at least get return value and 'this' types!");
-      }
-      var isClassMethodFunc = argTypes[1] !== null && classType !== null;
-      var needsDestructorStack = usesDestructorStack(argTypes);
-      var returns = argTypes[0].name !== 'void';
-      var expectedArgCount = argCount - 2;
-      var argsWired = new Array(expectedArgCount);
-      var invokerFuncArgs = [];
-      var destructors = [];
-      var invokerFn = function () {
-        if (arguments.length !== expectedArgCount) {
-          throwBindingError(
-            `function ${humanName} called with ${arguments.length} arguments, expected ${expectedArgCount}`,
-          );
-        }
-        destructors.length = 0;
-        var thisWired;
-        invokerFuncArgs.length = isClassMethodFunc ? 2 : 1;
-        invokerFuncArgs[0] = cppTargetFunc;
-        if (isClassMethodFunc) {
-          thisWired = argTypes[1]['toWireType'](destructors, this);
-          invokerFuncArgs[1] = thisWired;
-        }
-        for (var i = 0; i < expectedArgCount; ++i) {
-          argsWired[i] = argTypes[i + 2]['toWireType'](destructors, arguments[i]);
-          invokerFuncArgs.push(argsWired[i]);
-        }
-        var rv = cppInvokerFunc.apply(null, invokerFuncArgs);
-        function onDone(rv) {
-          if (needsDestructorStack) {
-            runDestructors(destructors);
-          } else {
-            for (var i = isClassMethodFunc ? 1 : 2; i < argTypes.length; i++) {
-              var param = i === 1 ? thisWired : argsWired[i - 2];
-              if (argTypes[i].destructorFunction !== null) {
-                argTypes[i].destructorFunction(param);
-              }
-            }
-          }
-          if (returns) {
-            return argTypes[0]['fromWireType'](rv);
-          }
-        }
-        return onDone(rv);
-      };
-      return createNamedFunction(humanName, invokerFn);
-    }
-
-    var __embind_register_class_constructor = (
-      rawClassType,
-      argCount,
-      rawArgTypesAddr,
-      invokerSignature,
-      invoker,
-      rawConstructor,
-    ) => {
-      var rawArgTypes = heap32VectorToArray(argCount, rawArgTypesAddr);
-      invoker = embind__requireFunction(invokerSignature, invoker);
-      whenDependentTypesAreResolved([], [rawClassType], function (classType) {
-        classType = classType[0];
-        var humanName = `constructor ${classType.name}`;
-        if (undefined === classType.registeredClass.constructor_body) {
-          classType.registeredClass.constructor_body = [];
-        }
-        if (undefined !== classType.registeredClass.constructor_body[argCount - 1]) {
-          throw new BindingError(
-            `Cannot register multiple constructors with identical number of parameters (${argCount - 1}) for class '${
-              classType.name
-            }'! Overload resolution is currently only performed using the parameter count, not actual type info!`,
-          );
-        }
-        classType.registeredClass.constructor_body[argCount - 1] = () => {
-          throwUnboundTypeError(`Cannot construct ${classType.name} due to unbound types`, rawArgTypes);
-        };
-        whenDependentTypesAreResolved([], rawArgTypes, (argTypes) => {
-          argTypes.splice(1, 0, null);
-          classType.registeredClass.constructor_body[argCount - 1] = craftInvokerFunction(
-            humanName,
-            argTypes,
-            null,
-            invoker,
-            rawConstructor,
-          );
-          return [];
-        });
-        return [];
-      });
-    };
-
-    var getFunctionName = (signature) => {
-      signature = signature.trim();
-      const argsIndex = signature.indexOf('(');
-      if (argsIndex !== -1) {
-        return signature.substr(0, argsIndex);
-      } else {
-        return signature;
-      }
-    };
-
-    var __embind_register_class_function = (
-      rawClassType,
-      methodName,
-      argCount,
-      rawArgTypesAddr,
-      invokerSignature,
-      rawInvoker,
-      context,
-      isPureVirtual,
-      isAsync,
-    ) => {
-      var rawArgTypes = heap32VectorToArray(argCount, rawArgTypesAddr);
-      methodName = readLatin1String(methodName);
-      methodName = getFunctionName(methodName);
-      rawInvoker = embind__requireFunction(invokerSignature, rawInvoker);
-      whenDependentTypesAreResolved([], [rawClassType], function (classType) {
-        classType = classType[0];
-        var humanName = `${classType.name}.${methodName}`;
-        if (methodName.startsWith('@@')) {
-          methodName = Symbol[methodName.substring(2)];
-        }
-        if (isPureVirtual) {
-          classType.registeredClass.pureVirtualFunctions.push(methodName);
-        }
-        function unboundTypesHandler() {
-          throwUnboundTypeError(`Cannot call ${humanName} due to unbound types`, rawArgTypes);
-        }
-        var proto = classType.registeredClass.instancePrototype;
-        var method = proto[methodName];
-        if (
-          undefined === method ||
-          (undefined === method.overloadTable &&
-            method.className !== classType.name &&
-            method.argCount === argCount - 2)
-        ) {
-          unboundTypesHandler.argCount = argCount - 2;
-          unboundTypesHandler.className = classType.name;
-          proto[methodName] = unboundTypesHandler;
-        } else {
-          ensureOverloadTable(proto, methodName, humanName);
-          proto[methodName].overloadTable[argCount - 2] = unboundTypesHandler;
-        }
-        whenDependentTypesAreResolved([], rawArgTypes, function (argTypes) {
-          var memberFunction = craftInvokerFunction(humanName, argTypes, classType, rawInvoker, context, isAsync);
-          if (undefined === proto[methodName].overloadTable) {
-            memberFunction.argCount = argCount - 2;
-            proto[methodName] = memberFunction;
-          } else {
-            proto[methodName].overloadTable[argCount - 2] = memberFunction;
-          }
-          return [];
-        });
-        return [];
-      });
-    };
-
-    function handleAllocatorInit() {
-      Object.assign(
-        HandleAllocator.prototype,
-        /** @lends {HandleAllocator.prototype} */ {
-          get(id) {
-            return this.allocated[id];
-          },
-          has(id) {
-            return this.allocated[id] !== undefined;
-          },
-          allocate(handle) {
-            var id = this.freelist.pop() || this.allocated.length;
-            this.allocated[id] = handle;
-            return id;
-          },
-          free(id) {
-            this.allocated[id] = undefined;
-            this.freelist.push(id);
-          },
-        },
-      );
-    }
-
-    /** @constructor */ function HandleAllocator() {
-      this.allocated = [undefined];
-      this.freelist = [];
-    }
-
-    var emval_handles = new HandleAllocator();
-
-    var __emval_decref = (handle) => {
-      if (handle >= emval_handles.reserved && 0 === --emval_handles.get(handle).refcount) {
-        emval_handles.free(handle);
-      }
-    };
-
-    var count_emval_handles = () => {
-      var count = 0;
-      for (var i = emval_handles.reserved; i < emval_handles.allocated.length; ++i) {
-        if (emval_handles.allocated[i] !== undefined) {
-          ++count;
-        }
-      }
-      return count;
-    };
-
-    var init_emval = () => {
-      emval_handles.allocated.push(
-        {
-          value: undefined,
-        },
-        {
-          value: null,
-        },
-        {
-          value: true,
-        },
-        {
-          value: false,
-        },
-      );
-      emval_handles.reserved = emval_handles.allocated.length;
-      Module['count_emval_handles'] = count_emval_handles;
-    };
-
-    var Emval = {
-      toValue: (handle) => {
-        if (!handle) {
-          throwBindingError('Cannot use deleted val. handle = ' + handle);
-        }
-        return emval_handles.get(handle).value;
-      },
-      toHandle: (value) => {
-        switch (value) {
-          case undefined:
-            return 1;
-
-          case null:
-            return 2;
-
-          case true:
-            return 3;
-
-          case false:
-            return 4;
-
-          default: {
-            return emval_handles.allocate({
-              refcount: 1,
-              value: value,
-            });
-          }
-        }
-      },
-    };
-
-    var __embind_register_emval = (rawType, name) => {
-      name = readLatin1String(name);
-      registerType(rawType, {
-        name: name,
-        fromWireType: (handle) => {
-          var rv = Emval.toValue(handle);
-          __emval_decref(handle);
-          return rv;
-        },
-        toWireType: (destructors, value) => Emval.toHandle(value),
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: simpleReadValueFromPointer,
-        destructorFunction: null,
-      });
-    };
-
-    var enumReadValueFromPointer = (name, width, signed) => {
-      switch (width) {
-        case 1:
-          return signed
-            ? function (pointer) {
-                return this['fromWireType'](HEAP8[pointer >> 0]);
-              }
-            : function (pointer) {
-                return this['fromWireType'](HEAPU8[pointer >> 0]);
-              };
-
-        case 2:
-          return signed
-            ? function (pointer) {
-                return this['fromWireType'](HEAP16[pointer >> 1]);
-              }
-            : function (pointer) {
-                return this['fromWireType'](HEAPU16[pointer >> 1]);
-              };
-
-        case 4:
-          return signed
-            ? function (pointer) {
-                return this['fromWireType'](HEAP32[pointer >> 2]);
-              }
-            : function (pointer) {
-                return this['fromWireType'](HEAPU32[pointer >> 2]);
-              };
-
-        default:
-          throw new TypeError(`invalid integer width (${width}): ${name}`);
-      }
-    };
-
-    /** @suppress {globalThis} */ var __embind_register_enum = (rawType, name, size, isSigned) => {
-      name = readLatin1String(name);
-      function ctor() {}
-      ctor.values = {};
-      registerType(rawType, {
-        name: name,
-        constructor: ctor,
-        fromWireType: function (c) {
-          return this.constructor.values[c];
-        },
-        toWireType: (destructors, c) => c.value,
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: enumReadValueFromPointer(name, size, isSigned),
-        destructorFunction: null,
-      });
-      exposePublicSymbol(name, ctor);
-    };
-
-    var requireRegisteredType = (rawType, humanName) => {
-      var impl = registeredTypes[rawType];
-      if (undefined === impl) {
-        throwBindingError(humanName + ' has unknown type ' + getTypeName(rawType));
-      }
-      return impl;
-    };
-
-    var __embind_register_enum_value = (rawEnumType, name, enumValue) => {
-      var enumType = requireRegisteredType(rawEnumType, 'enum');
-      name = readLatin1String(name);
-      var Enum = enumType.constructor;
-      var Value = Object.create(enumType.constructor.prototype, {
-        value: {
-          value: enumValue,
-        },
-        constructor: {
-          value: createNamedFunction(`${enumType.name}_${name}`, function () {}),
-        },
-      });
-      Enum.values[enumValue] = Value;
-      Enum[name] = Value;
-    };
-
-    var embindRepr = (v) => {
-      if (v === null) {
-        return 'null';
-      }
-      var t = typeof v;
-      if (t === 'object' || t === 'array' || t === 'function') {
-        return v.toString();
-      } else {
-        return '' + v;
-      }
-    };
-
-    var floatReadValueFromPointer = (name, width) => {
-      switch (width) {
-        case 4:
-          return function (pointer) {
-            return this['fromWireType'](HEAPF32[pointer >> 2]);
-          };
-
-        case 8:
-          return function (pointer) {
-            return this['fromWireType'](HEAPF64[pointer >> 3]);
-          };
-
-        default:
-          throw new TypeError(`invalid float width (${width}): ${name}`);
-      }
-    };
-
-    var __embind_register_float = (rawType, name, size) => {
-      name = readLatin1String(name);
-      registerType(rawType, {
-        name: name,
-        fromWireType: (value) => value,
-        toWireType: (destructors, value) => value,
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: floatReadValueFromPointer(name, size),
-        destructorFunction: null,
-      });
-    };
-
-    var integerReadValueFromPointer = (name, width, signed) => {
-      switch (width) {
-        case 1:
-          return signed ? (pointer) => HEAP8[pointer >> 0] : (pointer) => HEAPU8[pointer >> 0];
-
-        case 2:
-          return signed ? (pointer) => HEAP16[pointer >> 1] : (pointer) => HEAPU16[pointer >> 1];
-
-        case 4:
-          return signed ? (pointer) => HEAP32[pointer >> 2] : (pointer) => HEAPU32[pointer >> 2];
-
-        default:
-          throw new TypeError(`invalid integer width (${width}): ${name}`);
-      }
-    };
-
-    /** @suppress {globalThis} */ var __embind_register_integer = (primitiveType, name, size, minRange, maxRange) => {
-      name = readLatin1String(name);
-      if (maxRange === -1) {
-        maxRange = 4294967295;
-      }
-      var fromWireType = (value) => value;
-      if (minRange === 0) {
-        var bitshift = 32 - 8 * size;
-        fromWireType = (value) => (value << bitshift) >>> bitshift;
-      }
-      var isUnsignedType = name.includes('unsigned');
-      var checkAssertions = (value, toTypeName) => {};
-      var toWireType;
-      if (isUnsignedType) {
-        toWireType = function (destructors, value) {
-          checkAssertions(value, this.name);
-          return value >>> 0;
-        };
-      } else {
-        toWireType = function (destructors, value) {
-          checkAssertions(value, this.name);
-          return value;
-        };
-      }
-      registerType(primitiveType, {
-        name: name,
-        fromWireType: fromWireType,
-        toWireType: toWireType,
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: integerReadValueFromPointer(name, size, minRange !== 0),
-        destructorFunction: null,
-      });
-    };
-
-    var __embind_register_memory_view = (rawType, dataTypeIndex, name) => {
-      var typeMapping = [
-        Int8Array,
-        Uint8Array,
-        Int16Array,
-        Uint16Array,
-        Int32Array,
-        Uint32Array,
-        Float32Array,
-        Float64Array,
-      ];
-      var TA = typeMapping[dataTypeIndex];
-      function decodeMemoryView(handle) {
-        var size = HEAPU32[handle >> 2];
-        var data = HEAPU32[(handle + 4) >> 2];
-        return new TA(HEAP8.buffer, data, size);
-      }
-      name = readLatin1String(name);
-      registerType(
-        rawType,
-        {
-          name: name,
-          fromWireType: decodeMemoryView,
-          argPackAdvance: GenericWireTypeSize,
-          readValueFromPointer: decodeMemoryView,
-        },
-        {
-          ignoreDuplicateRegistrations: true,
-        },
-      );
-    };
-
-    var __embind_register_std_string = (rawType, name) => {
-      name = readLatin1String(name);
-      var stdStringIsUTF8 = name === 'std::string';
-      registerType(rawType, {
-        name: name,
-        fromWireType(value) {
-          var length = HEAPU32[value >> 2];
-          var payload = value + 4;
-          var str;
-          if (stdStringIsUTF8) {
-            var decodeStartPtr = payload;
-            for (var i = 0; i <= length; ++i) {
-              var currentBytePtr = payload + i;
-              if (i == length || HEAPU8[currentBytePtr] == 0) {
-                var maxRead = currentBytePtr - decodeStartPtr;
-                var stringSegment = UTF8ToString(decodeStartPtr, maxRead);
-                if (str === undefined) {
-                  str = stringSegment;
-                } else {
-                  str += String.fromCharCode(0);
-                  str += stringSegment;
-                }
-                decodeStartPtr = currentBytePtr + 1;
-              }
-            }
-          } else {
-            var a = new Array(length);
-            for (var i = 0; i < length; ++i) {
-              a[i] = String.fromCharCode(HEAPU8[payload + i]);
-            }
-            str = a.join('');
-          }
-          _free(value);
-          return str;
-        },
-        toWireType(destructors, value) {
-          if (value instanceof ArrayBuffer) {
-            value = new Uint8Array(value);
-          }
-          var length;
-          var valueIsOfTypeString = typeof value == 'string';
-          if (
-            !(
-              valueIsOfTypeString ||
-              value instanceof Uint8Array ||
-              value instanceof Uint8ClampedArray ||
-              value instanceof Int8Array
-            )
-          ) {
-            throwBindingError('Cannot pass non-string to std::string');
-          }
-          if (stdStringIsUTF8 && valueIsOfTypeString) {
-            length = lengthBytesUTF8(value);
-          } else {
-            length = value.length;
-          }
-          var base = _malloc(4 + length + 1);
-          var ptr = base + 4;
-          HEAPU32[base >> 2] = length;
-          if (stdStringIsUTF8 && valueIsOfTypeString) {
-            stringToUTF8(value, ptr, length + 1);
-          } else {
-            if (valueIsOfTypeString) {
-              for (var i = 0; i < length; ++i) {
-                var charCode = value.charCodeAt(i);
-                if (charCode > 255) {
-                  _free(ptr);
-                  throwBindingError('String has UTF-16 code units that do not fit in 8 bits');
-                }
-                HEAPU8[ptr + i] = charCode;
-              }
-            } else {
-              for (var i = 0; i < length; ++i) {
-                HEAPU8[ptr + i] = value[i];
-              }
-            }
-          }
-          if (destructors !== null) {
-            destructors.push(_free, base);
-          }
-          return base;
-        },
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: readPointer,
-        destructorFunction(ptr) {
-          _free(ptr);
-        },
-      });
-    };
-
-    var UTF16Decoder = typeof TextDecoder != 'undefined' ? new TextDecoder('utf-16le') : undefined;
-
-    var UTF16ToString = (ptr, maxBytesToRead) => {
-      var endPtr = ptr;
-      var idx = endPtr >> 1;
-      var maxIdx = idx + maxBytesToRead / 2;
-      while (!(idx >= maxIdx) && HEAPU16[idx]) ++idx;
-      endPtr = idx << 1;
-      if (endPtr - ptr > 32 && UTF16Decoder) return UTF16Decoder.decode(HEAPU8.subarray(ptr, endPtr));
-      var str = '';
-      for (var i = 0; !(i >= maxBytesToRead / 2); ++i) {
-        var codeUnit = HEAP16[(ptr + i * 2) >> 1];
-        if (codeUnit == 0) break;
-        str += String.fromCharCode(codeUnit);
-      }
-      return str;
-    };
-
-    var stringToUTF16 = (str, outPtr, maxBytesToWrite) => {
-      maxBytesToWrite ??= 2147483647;
-      if (maxBytesToWrite < 2) return 0;
-      maxBytesToWrite -= 2;
-      var startPtr = outPtr;
-      var numCharsToWrite = maxBytesToWrite < str.length * 2 ? maxBytesToWrite / 2 : str.length;
-      for (var i = 0; i < numCharsToWrite; ++i) {
-        var codeUnit = str.charCodeAt(i);
-        HEAP16[outPtr >> 1] = codeUnit;
-        outPtr += 2;
-      }
-      HEAP16[outPtr >> 1] = 0;
-      return outPtr - startPtr;
-    };
-
-    var lengthBytesUTF16 = (str) => str.length * 2;
-
-    var UTF32ToString = (ptr, maxBytesToRead) => {
-      var i = 0;
-      var str = '';
-      while (!(i >= maxBytesToRead / 4)) {
-        var utf32 = HEAP32[(ptr + i * 4) >> 2];
-        if (utf32 == 0) break;
-        ++i;
-        if (utf32 >= 65536) {
-          var ch = utf32 - 65536;
-          str += String.fromCharCode(55296 | (ch >> 10), 56320 | (ch & 1023));
-        } else {
-          str += String.fromCharCode(utf32);
-        }
-      }
-      return str;
-    };
-
-    var stringToUTF32 = (str, outPtr, maxBytesToWrite) => {
-      maxBytesToWrite ??= 2147483647;
-      if (maxBytesToWrite < 4) return 0;
-      var startPtr = outPtr;
-      var endPtr = startPtr + maxBytesToWrite - 4;
-      for (var i = 0; i < str.length; ++i) {
-        var codeUnit = str.charCodeAt(i);
-        if (codeUnit >= 55296 && codeUnit <= 57343) {
-          var trailSurrogate = str.charCodeAt(++i);
-          codeUnit = (65536 + ((codeUnit & 1023) << 10)) | (trailSurrogate & 1023);
-        }
-        HEAP32[outPtr >> 2] = codeUnit;
-        outPtr += 4;
-        if (outPtr + 4 > endPtr) break;
-      }
-      HEAP32[outPtr >> 2] = 0;
-      return outPtr - startPtr;
-    };
-
-    var lengthBytesUTF32 = (str) => {
-      var len = 0;
-      for (var i = 0; i < str.length; ++i) {
-        var codeUnit = str.charCodeAt(i);
-        if (codeUnit >= 55296 && codeUnit <= 57343) ++i;
-        len += 4;
-      }
-      return len;
-    };
-
-    var __embind_register_std_wstring = (rawType, charSize, name) => {
-      name = readLatin1String(name);
-      var decodeString, encodeString, getHeap, lengthBytesUTF, shift;
-      if (charSize === 2) {
-        decodeString = UTF16ToString;
-        encodeString = stringToUTF16;
-        lengthBytesUTF = lengthBytesUTF16;
-        getHeap = () => HEAPU16;
-        shift = 1;
-      } else if (charSize === 4) {
-        decodeString = UTF32ToString;
-        encodeString = stringToUTF32;
-        lengthBytesUTF = lengthBytesUTF32;
-        getHeap = () => HEAPU32;
-        shift = 2;
-      }
-      registerType(rawType, {
-        name: name,
-        fromWireType: (value) => {
-          var length = HEAPU32[value >> 2];
-          var HEAP = getHeap();
-          var str;
-          var decodeStartPtr = value + 4;
-          for (var i = 0; i <= length; ++i) {
-            var currentBytePtr = value + 4 + i * charSize;
-            if (i == length || HEAP[currentBytePtr >> shift] == 0) {
-              var maxReadBytes = currentBytePtr - decodeStartPtr;
-              var stringSegment = decodeString(decodeStartPtr, maxReadBytes);
-              if (str === undefined) {
-                str = stringSegment;
-              } else {
-                str += String.fromCharCode(0);
-                str += stringSegment;
-              }
-              decodeStartPtr = currentBytePtr + charSize;
-            }
-          }
-          _free(value);
-          return str;
-        },
-        toWireType: (destructors, value) => {
-          if (!(typeof value == 'string')) {
-            throwBindingError(`Cannot pass non-string to C++ string type ${name}`);
-          }
-          var length = lengthBytesUTF(value);
-          var ptr = _malloc(4 + length + charSize);
-          HEAPU32[ptr >> 2] = length >> shift;
-          encodeString(value, ptr + 4, length + charSize);
-          if (destructors !== null) {
-            destructors.push(_free, ptr);
-          }
-          return ptr;
-        },
-        argPackAdvance: GenericWireTypeSize,
-        readValueFromPointer: simpleReadValueFromPointer,
-        destructorFunction(ptr) {
-          _free(ptr);
-        },
-      });
-    };
-
-    var __embind_register_value_object = (
-      rawType,
-      name,
-      constructorSignature,
-      rawConstructor,
-      destructorSignature,
-      rawDestructor,
-    ) => {
-      structRegistrations[rawType] = {
-        name: readLatin1String(name),
-        rawConstructor: embind__requireFunction(constructorSignature, rawConstructor),
-        rawDestructor: embind__requireFunction(destructorSignature, rawDestructor),
-        fields: [],
-      };
-    };
-
-    var __embind_register_value_object_field = (
-      structType,
-      fieldName,
-      getterReturnType,
-      getterSignature,
-      getter,
-      getterContext,
-      setterArgumentType,
-      setterSignature,
-      setter,
-      setterContext,
-    ) => {
-      structRegistrations[structType].fields.push({
-        fieldName: readLatin1String(fieldName),
-        getterReturnType: getterReturnType,
-        getter: embind__requireFunction(getterSignature, getter),
-        getterContext: getterContext,
-        setterArgumentType: setterArgumentType,
-        setter: embind__requireFunction(setterSignature, setter),
-        setterContext: setterContext,
-      });
-    };
-
-    var __embind_register_void = (rawType, name) => {
-      name = readLatin1String(name);
-      registerType(rawType, {
-        isVoid: true,
-        name: name,
-        argPackAdvance: 0,
-        fromWireType: () => undefined,
-        toWireType: (destructors, o) => undefined,
-      });
-    };
-
-    var __emscripten_throw_longjmp = () => {
-      throw Infinity;
-    };
-
-    var __emval_incref = (handle) => {
-      if (handle > 4) {
-        emval_handles.get(handle).refcount += 1;
-      }
-    };
-
-    var __emval_take_value = (type, arg) => {
-      type = requireRegisteredType(type, '_emval_take_value');
-      var v = type['readValueFromPointer'](arg);
-      return Emval.toHandle(v);
-    };
-
-    var _abort = () => {
-      abort('');
-    };
-
-    var _emscripten_get_now;
-
-    _emscripten_get_now = () => performance.now();
-
-    var _emscripten_memcpy_js = (dest, src, num) => HEAPU8.copyWithin(dest, src, src + num);
-
-    var getHeapMax = () => 2147483648;
-
-    var growMemory = (size) => {
-      var b = wasmMemory.buffer;
-      var pages = (size - b.byteLength + 65535) / 65536;
-      try {
-        wasmMemory.grow(pages);
-        updateMemoryViews();
-        return 1;
-      } /*success*/ catch (e) {}
-    };
-
-    var _emscripten_resize_heap = (requestedSize) => {
-      var oldSize = HEAPU8.length;
-      requestedSize >>>= 0;
-      var maxHeapSize = getHeapMax();
-      if (requestedSize > maxHeapSize) {
-        return false;
-      }
-      var alignUp = (x, multiple) => x + ((multiple - (x % multiple)) % multiple);
-      for (var cutDown = 1; cutDown <= 4; cutDown *= 2) {
-        var overGrownHeapSize = oldSize * (1 + 0.2 / cutDown);
-        overGrownHeapSize = Math.min(overGrownHeapSize, requestedSize + 100663296);
-        var newSize = Math.min(maxHeapSize, alignUp(Math.max(requestedSize, overGrownHeapSize), 65536));
-        var replacement = growMemory(newSize);
-        if (replacement) {
-          return true;
-        }
-      }
-      return false;
-    };
-
-    var ENV = {};
-
-    var getExecutableName = () => thisProgram || './this.program';
-
-    var getEnvStrings = () => {
-      if (!getEnvStrings.strings) {
-        var lang =
-          ((typeof navigator == 'object' && navigator.languages && navigator.languages[0]) || 'C').replace('-', '_') +
-          '.UTF-8';
-        var env = {
-          USER: 'web_user',
-          LOGNAME: 'web_user',
-          PATH: '/',
-          PWD: '/',
-          HOME: '/home/web_user',
-          LANG: lang,
-          _: getExecutableName(),
-        };
-        for (var x in ENV) {
-          if (ENV[x] === undefined) delete env[x];
-          else env[x] = ENV[x];
-        }
-        var strings = [];
-        for (var x in env) {
-          strings.push(`${x}=${env[x]}`);
-        }
-        getEnvStrings.strings = strings;
-      }
-      return getEnvStrings.strings;
-    };
-
-    var stringToAscii = (str, buffer) => {
-      for (var i = 0; i < str.length; ++i) {
-        HEAP8[buffer++ >> 0] = str.charCodeAt(i);
-      }
-      HEAP8[buffer >> 0] = 0;
-    };
-
-    var _environ_get = (__environ, environ_buf) => {
-      var bufSize = 0;
-      getEnvStrings().forEach((string, i) => {
-        var ptr = environ_buf + bufSize;
-        HEAPU32[(__environ + i * 4) >> 2] = ptr;
-        stringToAscii(string, ptr);
-        bufSize += string.length + 1;
-      });
-      return 0;
-    };
-
-    var _environ_sizes_get = (penviron_count, penviron_buf_size) => {
-      var strings = getEnvStrings();
-      HEAPU32[penviron_count >> 2] = strings.length;
-      var bufSize = 0;
-      strings.forEach((string) => (bufSize += string.length + 1));
-      HEAPU32[penviron_buf_size >> 2] = bufSize;
-      return 0;
-    };
-
-    var _fd_close = (fd) => 52;
-
-    var _fd_read = (fd, iov, iovcnt, pnum) => 52;
-
-    var convertI32PairToI53Checked = (lo, hi) =>
-      (hi + 2097152) >>> 0 < 4194305 - !!lo ? (lo >>> 0) + hi * 4294967296 : NaN;
-
-    function _fd_seek(fd, offset_low, offset_high, whence, newOffset) {
-      var offset = convertI32PairToI53Checked(offset_low, offset_high);
-      return 70;
-    }
-
-    var printCharBuffers = [null, [], []];
-
-    var printChar = (stream, curr) => {
-      var buffer = printCharBuffers[stream];
-      if (curr === 0 || curr === 10) {
-        (stream === 1 ? out : err)(UTF8ArrayToString(buffer, 0));
-        buffer.length = 0;
-      } else {
-        buffer.push(curr);
-      }
-    };
-
-    var _fd_write = (fd, iov, iovcnt, pnum) => {
-      var num = 0;
-      for (var i = 0; i < iovcnt; i++) {
-        var ptr = HEAPU32[iov >> 2];
-        var len = HEAPU32[(iov + 4) >> 2];
-        iov += 8;
-        for (var j = 0; j < len; j++) {
-          printChar(fd, HEAPU8[ptr + j]);
-        }
-        num += len;
-      }
-      HEAPU32[pnum >> 2] = num;
-      return 0;
-    };
-
-    var initRandomFill = () => {
-      if (typeof crypto == 'object' && typeof crypto['getRandomValues'] == 'function') {
-        return (view) => crypto.getRandomValues(view);
-      } else abort('initRandomDevice');
-    };
-
-    var randomFill = (view) => (randomFill = initRandomFill())(view);
-
-    var _getentropy = (buffer, size) => {
-      randomFill(HEAPU8.subarray(buffer, buffer + size));
-      return 0;
-    };
-
-    var _llvm_eh_typeid_for = (type) => type;
-
-    var isLeapYear = (year) => year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
-
-    var arraySum = (array, index) => {
-      var sum = 0;
-      for (var i = 0; i <= index; sum += array[i++]) {}
-      return sum;
-    };
-
-    var MONTH_DAYS_LEAP = [31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
-    var MONTH_DAYS_REGULAR = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
-    var addDays = (date, days) => {
-      var newDate = new Date(date.getTime());
-      while (days > 0) {
-        var leap = isLeapYear(newDate.getFullYear());
-        var currentMonth = newDate.getMonth();
-        var daysInCurrentMonth = (leap ? MONTH_DAYS_LEAP : MONTH_DAYS_REGULAR)[currentMonth];
-        if (days > daysInCurrentMonth - newDate.getDate()) {
-          days -= daysInCurrentMonth - newDate.getDate() + 1;
-          newDate.setDate(1);
-          if (currentMonth < 11) {
-            newDate.setMonth(currentMonth + 1);
-          } else {
-            newDate.setMonth(0);
-            newDate.setFullYear(newDate.getFullYear() + 1);
-          }
-        } else {
-          newDate.setDate(newDate.getDate() + days);
-          return newDate;
-        }
-      }
-      return newDate;
-    };
-
-    /** @type {function(string, boolean=, number=)} */ function intArrayFromString(stringy, dontAddNull, length) {
-      var len = length > 0 ? length : lengthBytesUTF8(stringy) + 1;
-      var u8array = new Array(len);
-      var numBytesWritten = stringToUTF8Array(stringy, u8array, 0, u8array.length);
-      if (dontAddNull) u8array.length = numBytesWritten;
-      return u8array;
-    }
-
-    var writeArrayToMemory = (array, buffer) => {
-      HEAP8.set(array, buffer);
-    };
-
-    var _strftime = (s, maxsize, format, tm) => {
-      var tm_zone = HEAPU32[(tm + 40) >> 2];
-      var date = {
-        tm_sec: HEAP32[tm >> 2],
-        tm_min: HEAP32[(tm + 4) >> 2],
-        tm_hour: HEAP32[(tm + 8) >> 2],
-        tm_mday: HEAP32[(tm + 12) >> 2],
-        tm_mon: HEAP32[(tm + 16) >> 2],
-        tm_year: HEAP32[(tm + 20) >> 2],
-        tm_wday: HEAP32[(tm + 24) >> 2],
-        tm_yday: HEAP32[(tm + 28) >> 2],
-        tm_isdst: HEAP32[(tm + 32) >> 2],
-        tm_gmtoff: HEAP32[(tm + 36) >> 2],
-        tm_zone: tm_zone ? UTF8ToString(tm_zone) : '',
-      };
-      var pattern = UTF8ToString(format);
-      var EXPANSION_RULES_1 = {
-        '%c': '%a %b %d %H:%M:%S %Y',
-        '%D': '%m/%d/%y',
-        '%F': '%Y-%m-%d',
-        '%h': '%b',
-        '%r': '%I:%M:%S %p',
-        '%R': '%H:%M',
-        '%T': '%H:%M:%S',
-        '%x': '%m/%d/%y',
-        '%X': '%H:%M:%S',
-        '%Ec': '%c',
-        '%EC': '%C',
-        '%Ex': '%m/%d/%y',
-        '%EX': '%H:%M:%S',
-        '%Ey': '%y',
-        '%EY': '%Y',
-        '%Od': '%d',
-        '%Oe': '%e',
-        '%OH': '%H',
-        '%OI': '%I',
-        '%Om': '%m',
-        '%OM': '%M',
-        '%OS': '%S',
-        '%Ou': '%u',
-        '%OU': '%U',
-        '%OV': '%V',
-        '%Ow': '%w',
-        '%OW': '%W',
-        '%Oy': '%y',
-      };
-      for (var rule in EXPANSION_RULES_1) {
-        pattern = pattern.replace(new RegExp(rule, 'g'), EXPANSION_RULES_1[rule]);
-      }
-      var WEEKDAYS = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
-      var MONTHS = [
-        'January',
-        'February',
-        'March',
-        'April',
-        'May',
-        'June',
-        'July',
-        'August',
-        'September',
-        'October',
-        'November',
-        'December',
-      ];
-      function leadingSomething(value, digits, character) {
-        var str = typeof value == 'number' ? value.toString() : value || '';
-        while (str.length < digits) {
-          str = character[0] + str;
-        }
-        return str;
-      }
-      function leadingNulls(value, digits) {
-        return leadingSomething(value, digits, '0');
-      }
-      function compareByDay(date1, date2) {
-        function sgn(value) {
-          return value < 0 ? -1 : value > 0 ? 1 : 0;
-        }
-        var compare;
-        if ((compare = sgn(date1.getFullYear() - date2.getFullYear())) === 0) {
-          if ((compare = sgn(date1.getMonth() - date2.getMonth())) === 0) {
-            compare = sgn(date1.getDate() - date2.getDate());
-          }
-        }
-        return compare;
-      }
-      function getFirstWeekStartDate(janFourth) {
-        switch (janFourth.getDay()) {
-          case 0:
-            return new Date(janFourth.getFullYear() - 1, 11, 29);
-
-          case 1:
-            return janFourth;
-
-          case 2:
-            return new Date(janFourth.getFullYear(), 0, 3);
-
-          case 3:
-            return new Date(janFourth.getFullYear(), 0, 2);
-
-          case 4:
-            return new Date(janFourth.getFullYear(), 0, 1);
-
-          case 5:
-            return new Date(janFourth.getFullYear() - 1, 11, 31);
-
-          case 6:
-            return new Date(janFourth.getFullYear() - 1, 11, 30);
-        }
-      }
-      function getWeekBasedYear(date) {
-        var thisDate = addDays(new Date(date.tm_year + 1900, 0, 1), date.tm_yday);
-        var janFourthThisYear = new Date(thisDate.getFullYear(), 0, 4);
-        var janFourthNextYear = new Date(thisDate.getFullYear() + 1, 0, 4);
-        var firstWeekStartThisYear = getFirstWeekStartDate(janFourthThisYear);
-        var firstWeekStartNextYear = getFirstWeekStartDate(janFourthNextYear);
-        if (compareByDay(firstWeekStartThisYear, thisDate) <= 0) {
-          if (compareByDay(firstWeekStartNextYear, thisDate) <= 0) {
-            return thisDate.getFullYear() + 1;
-          }
-          return thisDate.getFullYear();
-        }
-        return thisDate.getFullYear() - 1;
-      }
-      var EXPANSION_RULES_2 = {
-        '%a': (date) => WEEKDAYS[date.tm_wday].substring(0, 3),
-        '%A': (date) => WEEKDAYS[date.tm_wday],
-        '%b': (date) => MONTHS[date.tm_mon].substring(0, 3),
-        '%B': (date) => MONTHS[date.tm_mon],
-        '%C': (date) => {
-          var year = date.tm_year + 1900;
-          return leadingNulls((year / 100) | 0, 2);
-        },
-        '%d': (date) => leadingNulls(date.tm_mday, 2),
-        '%e': (date) => leadingSomething(date.tm_mday, 2, ' '),
-        '%g': (date) => getWeekBasedYear(date).toString().substring(2),
-        '%G': (date) => getWeekBasedYear(date),
-        '%H': (date) => leadingNulls(date.tm_hour, 2),
-        '%I': (date) => {
-          var twelveHour = date.tm_hour;
-          if (twelveHour == 0) twelveHour = 12;
-          else if (twelveHour > 12) twelveHour -= 12;
-          return leadingNulls(twelveHour, 2);
-        },
-        '%j': (date) =>
-          leadingNulls(
-            date.tm_mday +
-              arraySum(isLeapYear(date.tm_year + 1900) ? MONTH_DAYS_LEAP : MONTH_DAYS_REGULAR, date.tm_mon - 1),
-            3,
-          ),
-        '%m': (date) => leadingNulls(date.tm_mon + 1, 2),
-        '%M': (date) => leadingNulls(date.tm_min, 2),
-        '%n': () => '\n',
-        '%p': (date) => {
-          if (date.tm_hour >= 0 && date.tm_hour < 12) {
-            return 'AM';
-          }
-          return 'PM';
-        },
-        '%S': (date) => leadingNulls(date.tm_sec, 2),
-        '%t': () => '\t',
-        '%u': (date) => date.tm_wday || 7,
-        '%U': (date) => {
-          var days = date.tm_yday + 7 - date.tm_wday;
-          return leadingNulls(Math.floor(days / 7), 2);
-        },
-        '%V': (date) => {
-          var val = Math.floor((date.tm_yday + 7 - ((date.tm_wday + 6) % 7)) / 7);
-          if ((date.tm_wday + 371 - date.tm_yday - 2) % 7 <= 2) {
-            val++;
-          }
-          if (!val) {
-            val = 52;
-            var dec31 = (date.tm_wday + 7 - date.tm_yday - 1) % 7;
-            if (dec31 == 4 || (dec31 == 5 && isLeapYear((date.tm_year % 400) - 1))) {
-              val++;
-            }
-          } else if (val == 53) {
-            var jan1 = (date.tm_wday + 371 - date.tm_yday) % 7;
-            if (jan1 != 4 && (jan1 != 3 || !isLeapYear(date.tm_year))) val = 1;
-          }
-          return leadingNulls(val, 2);
-        },
-        '%w': (date) => date.tm_wday,
-        '%W': (date) => {
-          var days = date.tm_yday + 7 - ((date.tm_wday + 6) % 7);
-          return leadingNulls(Math.floor(days / 7), 2);
-        },
-        '%y': (date) => (date.tm_year + 1900).toString().substring(2),
-        '%Y': (date) => date.tm_year + 1900,
-        '%z': (date) => {
-          var off = date.tm_gmtoff;
-          var ahead = off >= 0;
-          off = Math.abs(off) / 60;
-          off = (off / 60) * 100 + (off % 60);
-          return (ahead ? '+' : '-') + String('0000' + off).slice(-4);
-        },
-        '%Z': (date) => date.tm_zone,
-        '%%': () => '%',
-      };
-      pattern = pattern.replace(/%%/g, '\0\0');
-      for (var rule in EXPANSION_RULES_2) {
-        if (pattern.includes(rule)) {
-          pattern = pattern.replace(new RegExp(rule, 'g'), EXPANSION_RULES_2[rule](date));
-        }
-      }
-      pattern = pattern.replace(/\0\0/g, '%');
-      var bytes = intArrayFromString(pattern, false);
-      if (bytes.length > maxsize) {
-        return 0;
-      }
-      writeArrayToMemory(bytes, s);
-      return bytes.length - 1;
-    };
-
-    var _strftime_l = (s, maxsize, format, tm, loc) => _strftime(s, maxsize, format, tm);
-
-    InternalError = Module['InternalError'] = class InternalError extends Error {
-      constructor(message) {
-        super(message);
-        this.name = 'InternalError';
-      }
-    };
-
-    embind_init_charCodes();
-
-    BindingError = Module['BindingError'] = class BindingError extends Error {
-      constructor(message) {
-        super(message);
-        this.name = 'BindingError';
-      }
-    };
-
-    init_ClassHandle();
-
-    init_embind();
-
-    init_RegisteredPointer();
-
-    UnboundTypeError = Module['UnboundTypeError'] = extendError(Error, 'UnboundTypeError');
-
-    handleAllocatorInit();
-
-    init_emval();
-
-    var wasmImports = {
-      /** @export */ l: ___assert_fail,
-      /** @export */ na: ___cxa_begin_catch,
-      /** @export */ ma: ___cxa_end_catch,
-      /** @export */ a: ___cxa_find_matching_catch_2,
-      /** @export */ g: ___cxa_find_matching_catch_4,
-      /** @export */ w: ___cxa_throw,
-      /** @export */ c: ___resumeException,
-      /** @export */ E: ___syscall_fcntl64,
-      /** @export */ ba: ___syscall_getcwd,
-      /** @export */ da: ___syscall_ioctl,
-      /** @export */ ea: ___syscall_openat,
-      /** @export */ ra: __embind_finalize_value_object,
-      /** @export */ T: __embind_register_bigint,
-      /** @export */ ka: __embind_register_bool,
-      /** @export */ O: __embind_register_class,
-      /** @export */ M: __embind_register_class_constructor,
-      /** @export */ p: __embind_register_class_function,
-      /** @export */ ja: __embind_register_emval,
-      /** @export */ S: __embind_register_enum,
-      /** @export */ z: __embind_register_enum_value,
-      /** @export */ G: __embind_register_float,
-      /** @export */ s: __embind_register_integer,
-      /** @export */ o: __embind_register_memory_view,
-      /** @export */ H: __embind_register_std_string,
-      /** @export */ C: __embind_register_std_wstring,
-      /** @export */ P: __embind_register_value_object,
-      /** @export */ y: __embind_register_value_object_field,
-      /** @export */ la: __embind_register_void,
-      /** @export */ _: __emscripten_throw_longjmp,
-      /** @export */ ia: __emval_decref,
-      /** @export */ I: __emval_incref,
-      /** @export */ D: __emval_take_value,
-      /** @export */ L: _abort,
-      /** @export */ K: _emscripten_get_now,
-      /** @export */ fa: _emscripten_memcpy_js,
-      /** @export */ aa: _emscripten_resize_heap,
-      /** @export */ ga: _environ_get,
-      /** @export */ ha: _environ_sizes_get,
-      /** @export */ F: _fd_close,
-      /** @export */ ca: _fd_read,
-      /** @export */ R: _fd_seek,
-      /** @export */ B: _fd_write,
-      /** @export */ oa: _getentropy,
-      /** @export */ u: invoke_d,
-      /** @export */ A: invoke_fi,
-      /** @export */ N: invoke_i,
-      /** @export */ h: invoke_ii,
-      /** @export */ v: invoke_iif,
-      /** @export */ f: invoke_iii,
-      /** @export */ j: invoke_iiii,
-      /** @export */ m: invoke_iiiii,
-      /** @export */ r: invoke_iiiiii,
-      /** @export */ qa: invoke_iiiiiii,
-      /** @export */ J: invoke_iiiiiiii,
-      /** @export */ Z: invoke_ji,
-      /** @export */ Y: invoke_jii,
-      /** @export */ q: invoke_v,
-      /** @export */ b: invoke_vi,
-      /** @export */ t: invoke_vid,
-      /** @export */ d: invoke_vii,
-      /** @export */ pa: invoke_viiff,
-      /** @export */ e: invoke_viii,
-      /** @export */ k: invoke_viiii,
-      /** @export */ i: invoke_viiiii,
-      /** @export */ x: invoke_viiiiii,
-      /** @export */ Q: invoke_viiij,
-      /** @export */ V: invoke_viiiji,
-      /** @export */ W: invoke_viij,
-      /** @export */ U: invoke_viiji,
-      /** @export */ X: invoke_viijj,
-      /** @export */ n: _llvm_eh_typeid_for,
-      /** @export */ $: _strftime_l,
-    };
-
-    var wasmExports = createWasm();
-
-    var ___wasm_call_ctors = () => (___wasm_call_ctors = wasmExports['ta'])();
-
-    var _malloc = (a0) => (_malloc = wasmExports['ua'])(a0);
-
-    var _free = (a0) => (_free = wasmExports['wa'])(a0);
-
-    var ___errno_location = () => (___errno_location = wasmExports['__errno_location'])();
-
-    var ___getTypeName = (a0) => (___getTypeName = wasmExports['xa'])(a0);
-
-    var _htonl = (a0) => (_htonl = wasmExports['htonl'])(a0);
-
-    var _htons = (a0) => (_htons = wasmExports['htons'])(a0);
-
-    var _ntohs = (a0) => (_ntohs = wasmExports['ntohs'])(a0);
-
-    var _setThrew = (a0, a1) => (_setThrew = wasmExports['ya'])(a0, a1);
-
-    var setTempRet0 = (a0) => (setTempRet0 = wasmExports['za'])(a0);
-
-    var stackSave = () => (stackSave = wasmExports['Aa'])();
-
-    var stackRestore = (a0) => (stackRestore = wasmExports['Ba'])(a0);
-
-    var ___cxa_decrement_exception_refcount = (a0) => (___cxa_decrement_exception_refcount = wasmExports['Ca'])(a0);
-
-    var ___cxa_increment_exception_refcount = (a0) => (___cxa_increment_exception_refcount = wasmExports['Da'])(a0);
-
-    var ___cxa_can_catch = (a0, a1, a2) => (___cxa_can_catch = wasmExports['Ea'])(a0, a1, a2);
-
-    var ___cxa_is_pointer_type = (a0) => (___cxa_is_pointer_type = wasmExports['Fa'])(a0);
-
-    var dynCall_iijj = (Module['dynCall_iijj'] = (a0, a1, a2, a3, a4, a5) =>
-      (dynCall_iijj = Module['dynCall_iijj'] = wasmExports['Ga'])(a0, a1, a2, a3, a4, a5));
-
-    var dynCall_vijj = (Module['dynCall_vijj'] = (a0, a1, a2, a3, a4, a5) =>
-      (dynCall_vijj = Module['dynCall_vijj'] = wasmExports['Ha'])(a0, a1, a2, a3, a4, a5));
-
-    var dynCall_ji = (Module['dynCall_ji'] = (a0, a1) =>
-      (dynCall_ji = Module['dynCall_ji'] = wasmExports['Ia'])(a0, a1));
-
-    var dynCall_jii = (Module['dynCall_jii'] = (a0, a1, a2) =>
-      (dynCall_jii = Module['dynCall_jii'] = wasmExports['Ja'])(a0, a1, a2));
-
-    var dynCall_viijj = (Module['dynCall_viijj'] = (a0, a1, a2, a3, a4, a5, a6) =>
-      (dynCall_viijj = Module['dynCall_viijj'] = wasmExports['Ka'])(a0, a1, a2, a3, a4, a5, a6));
-
-    var dynCall_viij = (Module['dynCall_viij'] = (a0, a1, a2, a3, a4) =>
-      (dynCall_viij = Module['dynCall_viij'] = wasmExports['La'])(a0, a1, a2, a3, a4));
-
-    var dynCall_viiiji = (Module['dynCall_viiiji'] = (a0, a1, a2, a3, a4, a5, a6) =>
-      (dynCall_viiiji = Module['dynCall_viiiji'] = wasmExports['Ma'])(a0, a1, a2, a3, a4, a5, a6));
-
-    var dynCall_viiji = (Module['dynCall_viiji'] = (a0, a1, a2, a3, a4, a5) =>
-      (dynCall_viiji = Module['dynCall_viiji'] = wasmExports['Na'])(a0, a1, a2, a3, a4, a5));
-
-    var dynCall_viiij = (Module['dynCall_viiij'] = (a0, a1, a2, a3, a4, a5) =>
-      (dynCall_viiij = Module['dynCall_viiij'] = wasmExports['Oa'])(a0, a1, a2, a3, a4, a5));
-
-    var dynCall_jiji = (Module['dynCall_jiji'] = (a0, a1, a2, a3, a4) =>
-      (dynCall_jiji = Module['dynCall_jiji'] = wasmExports['Pa'])(a0, a1, a2, a3, a4));
-
-    var dynCall_viijii = (Module['dynCall_viijii'] = (a0, a1, a2, a3, a4, a5, a6) =>
-      (dynCall_viijii = Module['dynCall_viijii'] = wasmExports['Qa'])(a0, a1, a2, a3, a4, a5, a6));
-
-    var dynCall_iiiiij = (Module['dynCall_iiiiij'] = (a0, a1, a2, a3, a4, a5, a6) =>
-      (dynCall_iiiiij = Module['dynCall_iiiiij'] = wasmExports['Ra'])(a0, a1, a2, a3, a4, a5, a6));
-
-    var dynCall_iiiiijj = (Module['dynCall_iiiiijj'] = (a0, a1, a2, a3, a4, a5, a6, a7, a8) =>
-      (dynCall_iiiiijj = Module['dynCall_iiiiijj'] = wasmExports['Sa'])(a0, a1, a2, a3, a4, a5, a6, a7, a8));
-
-    var dynCall_iiiiiijj = (Module['dynCall_iiiiiijj'] = (a0, a1, a2, a3, a4, a5, a6, a7, a8, a9) =>
-      (dynCall_iiiiiijj = Module['dynCall_iiiiiijj'] = wasmExports['Ta'])(a0, a1, a2, a3, a4, a5, a6, a7, a8, a9));
-
-    function invoke_vi(index, a1) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iiii(index, a1, a2, a3) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2, a3);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viii(index, a1, a2, a3) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2, a3);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_vii(index, a1, a2) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiiii(index, a1, a2, a3, a4, a5) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2, a3, a4, a5);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_i(index) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)();
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_ii(index, a1) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_v(index) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)();
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiiiii(index, a1, a2, a3, a4, a5, a6) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2, a3, a4, a5, a6);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iiiii(index, a1, a2, a3, a4) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2, a3, a4);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iiiiii(index, a1, a2, a3, a4, a5) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2, a3, a4, a5);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_fi(index, a1) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iii(index, a1, a2) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iif(index, a1, a2) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_d(index) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)();
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_vid(index, a1, a2) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiii(index, a1, a2, a3, a4) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2, a3, a4);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iiiiiii(index, a1, a2, a3, a4, a5, a6) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2, a3, a4, a5, a6);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiff(index, a1, a2, a3, a4) {
-      var sp = stackSave();
-      try {
-        getWasmTableEntry(index)(a1, a2, a3, a4);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_iiiiiiii(index, a1, a2, a3, a4, a5, a6, a7) {
-      var sp = stackSave();
-      try {
-        return getWasmTableEntry(index)(a1, a2, a3, a4, a5, a6, a7);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_ji(index, a1) {
-      var sp = stackSave();
-      try {
-        return dynCall_ji(index, a1);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_jii(index, a1, a2) {
-      var sp = stackSave();
-      try {
-        return dynCall_jii(index, a1, a2);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viijj(index, a1, a2, a3, a4, a5, a6) {
-      var sp = stackSave();
-      try {
-        dynCall_viijj(index, a1, a2, a3, a4, a5, a6);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viij(index, a1, a2, a3, a4) {
-      var sp = stackSave();
-      try {
-        dynCall_viij(index, a1, a2, a3, a4);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiiji(index, a1, a2, a3, a4, a5, a6) {
-      var sp = stackSave();
-      try {
-        dynCall_viiiji(index, a1, a2, a3, a4, a5, a6);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiji(index, a1, a2, a3, a4, a5) {
-      var sp = stackSave();
-      try {
-        dynCall_viiji(index, a1, a2, a3, a4, a5);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    function invoke_viiij(index, a1, a2, a3, a4, a5) {
-      var sp = stackSave();
-      try {
-        dynCall_viiij(index, a1, a2, a3, a4, a5);
-      } catch (e) {
-        stackRestore(sp);
-        if (e !== e + 0) throw e;
-        _setThrew(1, 0);
-      }
-    }
-
-    var calledRun;
-
-    dependenciesFulfilled = function runCaller() {
-      if (!calledRun) run();
-      if (!calledRun) dependenciesFulfilled = runCaller;
-    };
-
-    function run() {
-      if (runDependencies > 0) {
-        return;
-      }
-      preRun();
-      if (runDependencies > 0) {
-        return;
-      }
-      function doRun() {
-        if (calledRun) return;
-        calledRun = true;
-        Module['calledRun'] = true;
-        if (ABORT) return;
-        initRuntime();
-        readyPromiseResolve(Module);
-        if (Module['onRuntimeInitialized']) Module['onRuntimeInitialized']();
-        postRun();
-      }
-      if (Module['setStatus']) {
-        Module['setStatus']('Running...');
-        setTimeout(function () {
-          setTimeout(function () {
-            Module['setStatus']('');
-          }, 1);
-          doRun();
-        }, 1);
-      } else {
-        doRun();
-      }
-    }
-
-    if (Module['preInit']) {
-      if (typeof Module['preInit'] == 'function') Module['preInit'] = [Module['preInit']];
-      while (Module['preInit'].length > 0) {
-        Module['preInit'].pop()();
-      }
-    }
-
-    run();
-
-    return moduleArg.ready;
-  };
-})();
-export default createDotLottiePlayerModule;
