+++ packages/web/src/core/index.ts
@@ -2,5 +2,5 @@
  * Copyright 2024 Design Barn Inc.
  */
 
-export type * from './dotlottie-player';
+export type * from './dotlottie-player.types';
 export * from './dotlottie-wasm-loader';
