+++ packages/web/vitest.config.ts
@@ -16,10 +16,10 @@ export default defineConfig({
       include: ['src/**/*.ts'],
       reporter: ['json', 'json-summary', 'text-summary', 'lcov'],
       thresholds: {
-        statements: 84,
-        branches: 76,
-        functions: 86,
-        lines: 85,
+        statements: 83,
+        branches: 75,
+        functions: 84,
+        lines: 86,
       },
     },
     testTimeout: 10000,
